refactor: restructure ecoNET300 constants into modular files for improved organization and maintainability
