--- custom_components/econet300/const.py
@@ -1,1824 +1,0 @@
-"""Constants for ecoNET300 integration.
-
-This module contains all constants organized by functionality:
-- Core integration constants
-- API endpoints and parameters
-- Device-specific sensor mappings
-- Entity configurations
-- Operation modes and status mappings
-- Mixer and thermostat configurations
-"""
-
-from typing import Any, Final
-
-from homeassistant.components.binary_sensor import BinarySensorDeviceClass
-from homeassistant.components.number import NumberDeviceClass
-from homeassistant.components.sensor import SensorDeviceClass, SensorStateClass
-from homeassistant.const import (
-    PERCENTAGE,
-    SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
-    STATE_OFF,
-    STATE_PAUSED,
-    STATE_PROBLEM,
-    STATE_UNKNOWN,
-    EntityCategory,
-    UnitOfEnergy,
-    UnitOfMass,
-    UnitOfPower,
-    UnitOfTemperature,
-    UnitOfTime,
-)
-
-# =============================================================================
-# CORE INTEGRATION CONSTANTS
-# =============================================================================
-DOMAIN = "econet300"
-SERVICE_API = "api"
-SERVICE_COORDINATOR = "coordinator"
-SERVICE_FUEL_SENSOR = "fuel_sensor"
-SERVICE_GET_SCHEDULE = "get_schedule"
-
-# regParams key for the fuel flow rate sensor (kg/h).
-SENSOR_FUEL_STREAM = "fuelStream"
-
-# Custom device class for fuel consumption meter to allow targeting with services
-DEVICE_CLASS_FUEL_METER = "econet300__fuel_meter"
-
-# Max sub-interval (seconds) for fuel consumption integration.
-# If fuelStream does not change within this period, integration is still triggered.
-FUEL_MAX_SUB_INTERVAL_SECONDS = 300
-
-# =============================================================================
-# COORDINATOR CONFIGURATION CONSTANTS
-# =============================================================================
-# Number of consecutive failures before creating a repair issue
-CONSECUTIVE_FAILURES_THRESHOLD = 5
-
-# Timeout in seconds for probing RM endpoint support (legacy-only modules return 404)
-RM_PROBE_TIMEOUT_SEC = 2
-
-# Max concurrent HTTP requests to the ecoNET module.
-# The module runs on a TP-Link MR3020 with very limited resources;
-# too many parallel connections cause timeouts (see GitHub issue #210).
-MAX_CONCURRENT_API_REQUESTS = 3
-
-# Coordinator update timeouts (seconds)
-# First update: static metadata cache is cold -> 6+ parallel RM API calls needed
-UPDATE_TIMEOUT_FIRST_SEC = 120
-# Subsequent updates: only dynamic data fetched (cached metadata reused)
-UPDATE_TIMEOUT_SEC = 30
-
-# RM endpoint dataset keys for data coordinator (order matches tasks list)
-RM_CORE_DATASET_KEYS = [
-    "currentDataParams",
-    "paramsNames",
-    "paramsData",
-    "langs",
-]
-
-RM_ADDITIONAL_DATASET_KEYS = [
-    "paramsDescs",
-    "paramsEnums",
-    "alarmsNames",
-]
-
-# =============================================================================
-# CACHE CONFIGURATION FOR STATIC METADATA
-# =============================================================================
-# Static metadata rarely changes - cache for 24 hours to reduce API load.
-CACHE_KEY_STATIC_METADATA = "static_metadata"
-CACHE_STATIC_METADATA_TTL = 86400  # 24 hours
-
-# =============================================================================
-# DEVICE INFORMATION CONSTANTS
-# =============================================================================
-DEVICE_INFO_MANUFACTURER = "PLUM"
-DEVICE_INFO_MODEL = "ecoNET300"
-DEVICE_INFO_CONTROLLER_NAME = "PLUM ecoNET300"
-DEVICE_INFO_MIXER_NAME = "Mixer device"
-DEVICE_INFO_LAMBDA_NAME = "Module Lambda"
-DEVICE_INFO_ECOSTER_NAME = "ecoSTER"
-DEVICE_INFO_HUW_NAME = "HUW Tank"
-DEVICE_INFO_BUFFER_NAME = "Buffer"
-DEVICE_INFO_SOLAR_NAME = "Solar"
-DEVICE_INFO_SERVICE_PARAMETERS_NAME = "Service Parameters"
-DEVICE_INFO_ADVANCED_PARAMETERS_NAME = "Advanced Parameters"
-
-CONF_ENTRY_TITLE = "ecoNET300"
-CONF_ENTRY_DESCRIPTION = "PLUM Econet300"
-
-# =============================================================================
-# HARDENING / POLLING
-# =============================================================================
-CONF_POLL_REG_PARAMS = "poll_reg_params"
-CONF_POLL_SYS_PARAMS = "poll_sys_params"
-CONF_POLL_EDIT_PARAMS = "poll_edit_params"
-
-# Conservative defaults for the small ecoNET300 web server.
-DEFAULT_POLL_REG_PARAMS = 15
-DEFAULT_POLL_SYS_PARAMS = 300
-DEFAULT_POLL_EDIT_PARAMS = 300
-
-# Keep last values during short glitches, but do not allow silent flat-lines forever.
-STALE_AFTER_SECONDS = 600
-
-
-# =============================================================================
-# DEVICE COMPONENT IDENTIFIERS
-# =============================================================================
-# These constants identify device components for grouping entities in Home Assistant.
-# When adding a new device type, add a constant here and update DEFAULT_COMPONENT_STATUS.
-COMPONENT_BOILER = "boiler"
-COMPONENT_HUW = "huw"
-COMPONENT_MIXER_1 = "mixer_1"
-COMPONENT_MIXER_2 = "mixer_2"
-COMPONENT_MIXER_3 = "mixer_3"
-COMPONENT_MIXER_4 = "mixer_4"
-COMPONENT_MIXER_5 = "mixer_5"
-COMPONENT_MIXER_6 = "mixer_6"
-COMPONENT_LAMBDA = "lambda"
-COMPONENT_BUFFER = "buffer"
-COMPONENT_SOLAR = "solar"
-
-# Default component status template - used when no reg_params available
-DEFAULT_COMPONENT_STATUS: dict[str, bool] = {
-    COMPONENT_BOILER: True,
-    COMPONENT_HUW: False,
-    COMPONENT_MIXER_1: False,
-    COMPONENT_MIXER_2: False,
-    COMPONENT_MIXER_3: False,
-    COMPONENT_MIXER_4: False,
-    COMPONENT_MIXER_5: False,
-    COMPONENT_MIXER_6: False,
-    COMPONENT_LAMBDA: False,
-    COMPONENT_BUFFER: False,
-    COMPONENT_SOLAR: False,
-}
-
-# =============================================================================
-# API ENDPOINT CONSTANTS
-# =============================================================================
-# endpoint url sysParams
-API_SYS_PARAMS_URI = "sysParams"
-
-# sysParams property names
-API_SYS_PARAMS_PARAM_UID = "uid"
-API_SYS_PARAMS_PARAM_MODEL_ID = "controllerID"
-API_SYS_PARAMS_PARAM_SW_REV = "softVer"
-API_SYS_PARAMS_PARAM_HW_VER = "routerType"
-
-#  endpoint url regParams
-API_REG_PARAMS_URI = "regParams"
-
-# regParams property names
-API_REG_PARAMS_PARAM_DATA = "curr"
-
-# endpoint url regParamsData
-API_REG_PARAMS_DATA_URI = "regParamsData"
-API_REG_PARAMS_DATA_PARAM_DATA = "data"
-
-# Legacy parameter modification endpoint (non-RM, uses newParamName)
-API_NEW_PARAM_URI = "newParam"
-
-# Editable parameters
-API_EDITABLE_PARAMS_LIMITS_URI = "rmCurrentDataParamsEdits"
-API_EDITABLE_PARAMS_LIMITS_DATA = "data"
-
-# Optional LAN endpoint for editable-parameter metadata (module-dependent; may 404)
-API_EDIT_PARAMS_URI = "editParams"
-
-# =============================================================================
-# RM... ENDPOINT CONSTANTS (Remote Menu API)
-# =============================================================================
-# These endpoints provide structured data for the ecoNET24 web interface
-# Based on analysis of dev_set1.js and test fixtures
-
-# Core RM endpoints for parameter management
-API_RM_PARAMS_NAMES_URI = "rmParamsNames"
-API_RM_PARAMS_DATA_URI = "rmParamsData"
-API_RM_PARAMS_DESCS_URI = "rmParamsDescs"
-API_RM_PARAMS_ENUMS_URI = "rmParamsEnums"
-API_RM_PARAMS_UNITS_NAMES_URI = "rmParamsUnitsNames"
-
-# RM endpoints for categories and structure
-API_RM_STRUCTURE_URI = "rmStructure"
-
-# RM endpoints for current data
-API_RM_CURRENT_DATA_PARAMS_URI = "rmCurrentDataParams"
-API_RM_CURRENT_DATA_PARAMS_EDITS_URI = "rmCurrentDataParamsEdits"
-
-# RM endpoints for system information
-API_RM_LANGS_URI = "rmLangs"
-API_RM_EXISTING_LANGS_URI = "rmExistingLangs"
-API_RM_LOCKS_NAMES_URI = "rmLocksNames"
-API_RM_ALARMS_NAMES_URI = "rmAlarmsNames"
-
-# RM endpoints for authentication and parameter modification
-API_RM_ACCESS_URI = "rmAccess"  # Service password authentication
-API_RM_NEW_PARAM_URI = "rmNewParam"  # Save parameter by index
-API_RM_CURR_NEW_PARAM_URI = "rmCurrNewParam"  # Save current param by key
-
-# RM endpoint data key (all endpoints use "data" as the key)
-API_RM_DATA_KEY = "data"
-
-# =============================================================================
-# RM STRUCTURE ENTRY TYPES
-# =============================================================================
-# These constants define the entry types in rmStructure API response
-# Each entry in the structure has a "type" field indicating its purpose
-
-RM_STRUCTURE_TYPE_CATEGORY = 0  # Category entry - defines category context
-RM_STRUCTURE_TYPE_PARAMETER = 1  # Editable parameter entry
-RM_STRUCTURE_TYPE_DATA_REF = 3  # Data reference (read-only, has data_id)
-RM_STRUCTURE_TYPE_MENU_GROUP = 7  # Menu group/header (resets pass_index)
-
-# =============================================================================
-# ALARM CODE CONSTANTS
-# =============================================================================
-ALARM_CODE_POWER_OUTAGE = 0
-ALARM_CODE_CONTINUES = 255
-
-# =============================================================================
-# OPERATION MODES AND STATUS MAPPINGS
-# =============================================================================
-OPERATION_MODE_NAMES = {
-    0: STATE_OFF,
-    1: "fire_up",
-    2: "operation",
-    3: "work",
-    4: "supervision",
-    5: STATE_PAUSED,  # "halted",
-    6: "stop",
-    7: "burning_off",
-    8: "manual",
-    9: STATE_PROBLEM,  # "alarm",
-    10: "unsealing",
-    11: "chimney",
-    12: "stabilization",
-    13: "no_transmission",
-}
-
-# =============================================================================
-# MIXER CONFIGURATION CONSTANTS
-# =============================================================================
-NUMBER_OF_AVAILABLE_MIXERS = 6  # Supports up to 6 mixers (ecoMAX850R2-X has 5)
-NUMBER_OF_AVAILABLE_ECOSTERS = 8  # Supports up to 8 ecoSTER thermostats
-MIXER_AVAILABILITY_KEY = "mixerTemp"
-MIXER_SET_AVAILABILITY_KEY = "mixerSetTemp"
-
-# Dynamically generate SENSOR_MIXER_KEY
-SENSOR_MIXER_KEY = {
-    i: {f"{MIXER_AVAILABILITY_KEY}{i}", f"{MIXER_SET_AVAILABILITY_KEY}{i}"}
-    for i in range(1, NUMBER_OF_AVAILABLE_MIXERS + 1)
-}
-
-# Mixer pump binary sensor keys
-MIXER_PUMP_BINARY_SENSOR_KEYS = {
-    f"mixerPumpWorks{i}" for i in range(1, NUMBER_OF_AVAILABLE_MIXERS + 1)
-}
-
-# Keywords that indicate mixer-related parameters for duplicate entity filtering
-MIXER_RELATED_KEYWORDS: list[str] = [
-    "mixer",
-    "valve",
-    "heating circuit",
-    "actuator",
-    "circuit",
-]
-
-# =============================================================================
-# DEVICE-SPECIFIC SENSOR MAPPINGS
-# =============================================================================
-# ecoMAX360i specific sensors (regParams ``curr`` + common sysParams diagnostics)
-ECOMAX360I_SENSORS = {
-    "PS",
-    "Circuit2thermostatTemp",
-    "TempClutch",
-    "Circuit3thermostatTemp",
-    "Circuit4thermostatTemp",
-    "Circuit5thermostatTemp",
-    "Circuit6thermostatTemp",
-    "Circuit7thermostatTemp",
-    "TempWthr",
-    "TempCircuit2",
-    "TempCircuit3",
-    "TempCircuit4",
-    "TempCircuit5",
-    "TempCircuit6",
-    "TempCircuit7",
-    "TempBuforUp",
-    "TempCWU",
-    "TempBuforDown",
-    "heatingUpperTemp",
-    "Circuit1thermostat",
-    "heating_work_state_pump4",
-    # 3-way valve and demand sensors
-    "flapValveStates",
-    "HeatDemanded",
-    "WaterPumpRunning",
-    # Heat pump (Axen) sensors
-    "AxenOutdoorTemp",
-    "AxenOutgoingTemp",
-    "AxenReturnTemp",
-    "AxenCompressorFreq",
-    "AxenUpperPump",
-    "AxenWorkState",
-    "HeatSourceCalcPresetTemp",
-    # Heat pump compressor temperatures
-    "afterCompressorTemp",
-    "beforeCompressorTemp",
-    "exhaustGasTemp",
-    "outdoorTemp",
-    # Heat pump status sensors
-    "HPStatusWorkMode",
-    "HPStatusPresetTemp",
-    "HPStatusControl",
-    # SSA (weather compensation) sensors
-    "ssaState",
-    "ssaCorr",
-    "ssaPrevTemp",
-    # Circuit comfort/eco temperature setpoints (read-only from regParams)
-    "Circuit1ComfortTemp",
-    "Circuit1EcoTemp",
-    "BufferTargetTemp",
-    "Circuit2ComfortTemp",
-    "Circuit2EcoTemp",
-    "Circuit3ComfortTemp",
-    "Circuit3EcoTemp",
-    "Circuit4ComfortTemp",
-    "Circuit4EcoTemp",
-    "Circuit5ComfortTemp",
-    "Circuit5EcoTemp",
-    "Circuit6ComfortTemp",
-    "Circuit6EcoTemp",
-    "Circuit7ComfortTemp",
-    "Circuit7EcoTemp",
-    # Phoenix / extended heat-pump sensors (decomposed from sensor.py EXTRA_SENSORS)
-    "BuforCalcSetTemp",
-    "Circuit1CalcTemp",
-    "Circuit2CalcTemp",
-    "Circuit4CalcTemp",
-    "Circuit5CalcTemp",
-    "Circuit6CalcTemp",
-    "Circuit7CalcTemp",
-    "HDWsetpointcalculate",
-    "CoolingPower",
-    "HeatingPower",
-    "ElectricPower",
-    "currentFlow",
-    "HPStatusBuffHeatStat",
-    "HPStatusCircPStat0",
-    "HPStatusCircPStat1",
-    "HPStatusCircPStat2",
-    "HPStatusCircPStat3",
-    "HPStatusCircPStat4",
-    "HPStatusCircPStat5",
-    "HPStatusCircPStat6",
-    "HPStatusFlowHeatStat",
-    "HPStatusHdwHeatStat",
-    "HPStatusUhsStat",
-    "HP_work_state_set_pump2",
-    "HP_work_state_set_pump3",
-    "HP_work_state_set_pump4",
-    "HP_work_state_set_pump5",
-    "PHNXcoil_temp",
-    "PHNXdischarge_temp",
-    "PHNXinletTemp",
-    "PHNXinletTemp_pump",
-    "PHNXinletTemp_pump2",
-    "PHNXinletTemp_pump3",
-    "PHNXinletTemp_pump5",
-    "PHNXoutletTemp",
-    "PHNXoutletTemp_pump2",
-    "PHNXoutletTemp_pump3",
-    "PHNXoutletTemp_pump4",
-    "PHNXoutletTemp_pump5",
-    "PHNXreg2045",
-    "PHNXreg2046",
-    "PHNXreg2071",
-    "PHNXreg2074",
-    "PHNXsuctionTemp",
-    # informationParams sensors (from editParams endpoint)
-    "TargetFlowTemp",
-    "ActualFlowTemp",
-    "ActualReturnTemp",
-    "FanSpeed",
-    "HeatPumpAmbient",
-    "ActualDHWTemp",
-    "Circuit1DesiredLWT",
-    "ElectricalPower",
-    "ThermalPower",
-    "COP",
-    "SCOP",
-    "FlowRate",
-    # editParams.data sensors
-    "AXENREGISTER64",
-    "AXENREGISTER65",
-    # sysParams diagnostics (same as DEFAULT_SENSORS subset)
-    "controllerID",
-    "moduleASoftVer",
-    "moduleBSoftVer",
-    "moduleCSoftVer",
-    "moduleLambdaSoftVer",
-    "modulePanelSoftVer",
-    "moduleEcoSTERSoftVer",
-    "protocolType",
-    "quality",
-    "routerType",
-    "signal",
-    "softVer",
-}
-
-# informationParams: sensor key -> parameter ID in editParams.informationParams
-# These are read-only heat pump / status sensors for ecoMAX360i.
-INFORMATION_PARAMS_SENSOR_MAP: dict[str, str] = {
-    "TargetFlowTemp": "12",
-    "ActualFlowTemp": "14",
-    "ActualReturnTemp": "15",
-    "FanSpeed": "22",
-    "HeatPumpAmbient": "23",
-    "HeatDemanded": "26",
-    "ActualDHWTemp": "61",
-    "Circuit1DesiredLWT": "93",
-    "ElectricalPower": "211",
-    "ThermalPower": "212",
-    "COP": "221",
-    "SCOP": "222",
-    "FlowRate": "231",
-}
-
-# editParams.data: sensor key -> parameter ID in editParams["data"]
-# These are read-only Axen register sensors for ecoMAX360i.
-EDIT_PARAMS_DATA_SENSOR_MAP: dict[str, str] = {
-    "AXENREGISTER64": "1211",
-    "AXENREGISTER65": "1212",
-}
-
-# ecoSTER thermostat sensors (if moduleEcoSTERSoftVer is not None)
-ECOSTER_SENSORS = {
-    # ecoSTER temperature sensors
-    "ecoSterTemp1",
-    "ecoSterTemp2",
-    "ecoSterTemp3",
-    "ecoSterTemp4",
-    "ecoSterTemp5",
-    "ecoSterTemp6",
-    "ecoSterTemp7",
-    "ecoSterTemp8",
-    # ecoSTER setpoint sensors
-    "ecoSterSetTemp1",
-    "ecoSterSetTemp2",
-    "ecoSterSetTemp3",
-    "ecoSterSetTemp4",
-    "ecoSterSetTemp5",
-    "ecoSterSetTemp6",
-    "ecoSterSetTemp7",
-    "ecoSterSetTemp8",
-    # ecoSTER mode sensors
-    "ecoSterMode1",
-    "ecoSterMode2",
-    "ecoSterMode3",
-    "ecoSterMode4",
-    "ecoSterMode5",
-    "ecoSterMode6",
-    "ecoSterMode7",
-    "ecoSterMode8",
-}
-
-# Lambda sensor module
-LAMBDA_SENSORS = {
-    "lambdaStatus",
-    "lambdaSet",
-    "lambdaLevel",
-}
-
-# ecoSOL solar collector sensors (all ecoSOL [n] models with matching regParams)
-ECOSOL_SENSORS = {
-    # Temperature sensors
-    "T1",  # Collector temperature
-    "T2",  # Tank temperature
-    "T3",  # Tank temperature
-    "T4",  # Return temperature
-    "T5",  # Collector temperature - power measurement
-    "T6",  # Temperature sensor
-    "TzCWU",  # Hot water temperature
-    # Pump status sensors
-    "P1",  # Pump 1 status
-    "P2",  # Pump 2 status
-    # Output status
-    "H",  # Output status
-    # Power / energy (regParams; ecoSOL 301 flat JSON uses subset)
-    "Moc_chwilowa",  # Instantaneous solar collector power (kW)
-    "Uzysk_ca_kowity",  # Cumulative solar heat yield (kWh)
-    # Diagnostic sensors
-    "ecosrvAddr",
-    "quality",
-    "signal",
-    "softVer",
-    "routerType",
-    "protocolType",
-    "controllerID",
-    "ecosrvSoftVer",
-}
-
-# Reference-only dict key for SENSOR_MAP_KEY / BINARY_SENSOR_MAP_KEY (not a live controllerID)
-ECOSOL_CONTROLLER_MAP_REFERENCE_KEY = "ecoSOL [n]"
-
-# Default sensors for most controllers
-DEFAULT_SENSORS = {
-    "boilerPower",
-    "boilerPowerKW",
-    "tempFeeder",
-    "fuelLevel",
-    "tempCO",
-    "tempCOSet",
-    "statusCWU",
-    "tempCWU",
-    "tempCWUSet",
-    "tempFlueGas",
-    "mode",
-    "fanPower",
-    "thermostat",
-    "tempExternalSensor",
-    "tempLowerBuffer",
-    "tempUpperBuffer",
-    "quality",
-    "signal",
-    "softVer",
-    "controllerID",
-    "moduleASoftVer",
-    "moduleBSoftVer",
-    "moduleCSoftVer",
-    "moduleLambdaSoftVer",
-    "modulePanelSoftVer",
-    "moduleEcoSTERSoftVer",
-    # ecoMAX850R2-X specific sensors
-    "fuelConsum",
-    SENSOR_FUEL_STREAM,
-    "tempBack",
-    "transmission",
-    "statusCO",
-    # Diagnostic sensors
-    "routerType",
-    "protocolType",
-}
-
-# Main sensor mapping by controller type
-# Sensor platform: is_ecosol_controller() → ECOSOL_SENSORS;
-# is_ecomax360i_controller() → ECOMAX360I_SENSORS (see sensor.py); else DEFAULT_SENSORS.
-# ecoSter / lambda entries remain reference-only for now.
-SENSOR_MAP_KEY = {
-    "ecoMAX360i": ECOMAX360I_SENSORS,  # Runtime via is_ecomax360i_controller()
-    ECOSOL_CONTROLLER_MAP_REFERENCE_KEY: ECOSOL_SENSORS,  # Reference — runtime uses is_ecosol_controller()
-    "ecoSter": ECOSTER_SENSORS,  # Reference only - not used
-    COMPONENT_LAMBDA: LAMBDA_SENSORS,  # Reference only - not used
-    "_default": DEFAULT_SENSORS,  # Default for non-ecoSOL controllers
-}
-
-# =============================================================================
-# DEVICE-SPECIFIC BINARY SENSOR MAPPINGS
-# =============================================================================
-# Default binary sensors for most controllers
-DEFAULT_BINARY_SENSORS = {
-    "lighterWorks",
-    "pumpCOWorks",
-    "fanWorks",
-    "feederWorks",
-    "pumpFireplaceWorks",
-    "pumpCWUWorks",
-    "mainSrv",
-    "wifi",
-    "lan",
-    "fuelConsumptionCalc",
-    "ecosrvHttps",
-    # ecoMAX850R2-X specific binary sensors
-    "contactGZC",
-    "contactGZCActive",
-    "pumpCirculationWorks",
-    "pumpSolarWorks",
-}
-
-# ecoSTER thermostat binary sensors
-ECOSTER_BINARY_SENSORS = {
-    # ecoSTER contact sensors
-    "ecoSterContacts1",
-    "ecoSterContacts2",
-    "ecoSterContacts3",
-    "ecoSterContacts4",
-    "ecoSterContacts5",
-    "ecoSterContacts6",
-    "ecoSterContacts7",
-    "ecoSterContacts8",
-    # ecoSTER day schedule sensors
-    "ecoSterDaySched1",
-    "ecoSterDaySched2",
-    "ecoSterDaySched3",
-    "ecoSterDaySched4",
-    "ecoSterDaySched5",
-    "ecoSterDaySched6",
-    "ecoSterDaySched7",
-    "ecoSterDaySched8",
-}
-
-# ecoSOL solar collector binary sensors (all ecoSOL [n] models)
-ECOSOL_BINARY_SENSORS = {
-    "wifi",
-    "lan",
-    "mainSrv",
-    "fuelConsumptionCalc",
-    "ecosrvHttps",
-}
-
-# Main binary sensor mapping by controller type
-# All controllers use DEFAULT_BINARY_SENSORS (specific mappings are for reference only)
-BINARY_SENSOR_MAP_KEY = {
-    "_default": DEFAULT_BINARY_SENSORS,  # Always used for all controllers
-    ECOSOL_CONTROLLER_MAP_REFERENCE_KEY: ECOSOL_BINARY_SENSORS,  # Reference — runtime uses is_ecosol_controller()
-    "ecoSter": ECOSTER_BINARY_SENSORS,  # Reference only - not used
-}
-
-# =============================================================================
-# NUMBER ENTITY MAPPINGS
-# =============================================================================
-NUMBER_MAP = {
-    "1280": "tempCOSet",  # Boiler temperature setpoint
-    "1281": "tempCWUSet",  # Hot water temperature setpoint
-    "1287": "mixerSetTemp1",  # Mixer 1 temperature setpoint
-    "1288": "mixerSetTemp2",  # Mixer 2 temperature setpoint
-    "1289": "mixerSetTemp3",  # Mixer 3 temperature setpoint
-    "1290": "mixerSetTemp4",  # Mixer 4 temperature setpoint
-    "1291": "mixerSetTemp5",  # Mixer 5 temperature setpoint
-    "1292": "mixerSetTemp6",  # Mixer 6 temperature setpoint
-    "55": "heaterMode",  # Heater mode (Summer/Winter/Auto)
-}
-
-# =============================================================================
-# SELECT ENTITY MAPPINGS
-# =============================================================================
-
-# Select entity parameter mappings (for writing settings)
-SELECT_KEY_POST_INDEX: dict[str, str] = {
-    "heaterMode": "55",  # Heater mode winter/summer/auto
-    # Add more select entities here as they are created
-    # "pumpMode": "56",
-    # "fanMode": "57",
-}
-
-# Select entity current state mappings (for reading current state)
-SELECT_KEY_GET_INDEX: dict[str, str] = {
-    "heaterMode": "2049",  # Heater mode current state
-    # Add more select entities here as they are created
-    # "pumpMode": "2050",
-    # "fanMode": "2051",
-}
-
-# Select entity value mappings (numeric value -> display name)
-SELECT_KEY_VALUES: dict[str, dict[int, str]] = {
-    "heaterMode": {
-        0: "winter",
-        1: "summer",
-        2: "auto",
-    },
-    # Add more select entities here as they are created
-    # "pumpMode": {
-    #     0: "Off",
-    #     1: "On",
-    #     2: "Auto",
-    # },
-}
-
-
-# Legacy constants for backward compatibility
-HEATER_MODE_VALUES = SELECT_KEY_VALUES["heaterMode"]
-HEATER_MODE_PARAM_INDEX = SELECT_KEY_POST_INDEX["heaterMode"]
-HEATER_MODE_CURRENT_STATE_PARAM = SELECT_KEY_GET_INDEX["heaterMode"]
-
-# =============================================================================
-# PARAMETER ENDPOINT MAPPINGS
-# =============================================================================
-# Parameters that use the rmNewParam endpoint with newParamIndex
-RMNEWPARAM_PARAMS = {
-    "55",  # Heater mode (Summer/Winter/Auto)
-    # Add other parameters here that need rmNewParam endpoint
-    # Example: "56", "57", etc.
-}
-
-# Control parameters that use the newParam endpoint with newParamName
-CONTROL_PARAMS = {
-    "BOILER_CONTROL",  # Boiler ON/OFF control
-    # Add other control parameters here
-    # Example: "PUMP_CONTROL", "FAN_CONTROL", etc.
-}
-
-# Individual control parameter constants
-BOILER_CONTROL = "BOILER_CONTROL"
-
-# =============================================================================
-# DYNAMIC ENTITY UNIT MAPPINGS
-# =============================================================================
-# Mapping from ecoNET unit names to Home Assistant units
-# Used for dynamic number entity creation from merged parameter data
-UNIT_NAME_TO_HA_UNIT = {
-    "%": PERCENTAGE,
-    "°C": UnitOfTemperature.CELSIUS,
-    "sek.": UnitOfTime.SECONDS,
-    "min.": UnitOfTime.MINUTES,
-    "h.": UnitOfTime.HOURS,
-    "kg": UnitOfMass.KILOGRAMS,
-    "kg/h": "kg/h",  # Mass flow rate for fuel stream
-    "kW": UnitOfPower.KILO_WATT,
-    "kWh": UnitOfEnergy.KILO_WATT_HOUR,
-    "r/min": "r/min",  # Custom unit for revolutions per minute
-}
-
-# =============================================================================
-# UNIT MAPPINGS (shared by dynamic entity helpers)
-# =============================================================================
-# Maps unit index from rmCurrentDataParams to human-readable unit string.
-UNIT_INDEX_TO_NAME: dict[int, str] = {
-    0: "",
-    1: "°C",
-    2: "sek.",
-    3: "min.",
-    4: "h.",
-    5: "%",
-    6: "kg",
-    7: "kW",
-    8: "r/min",
-    31: "",  # Boolean/state indicator (no unit)
-}
-
-# --- Static entity deduplication ---------------------------------------------
-# regParamsData IDs already handled by static entities (number, select).
-# These are skipped in the Options Flow key listing to avoid duplicates.
-STATIC_REGPARAMS_DATA_IDS: set[str] = (
-    set(NUMBER_MAP.keys())
-    | set(SELECT_KEY_POST_INDEX.values())
-    | set(SELECT_KEY_GET_INDEX.values())
-)
-
-# All static regParams keys (sensors + binary sensors across all controllers).
-# Used to filter already-mapped keys in the Options Flow when selecting
-# custom entities from the regParams endpoint.
-STATIC_REGPARAMS_KEYS: set[str] = (
-    DEFAULT_SENSORS
-    | DEFAULT_BINARY_SENSORS
-    | ECOMAX360I_SENSORS
-    | ECOSTER_SENSORS
-    | LAMBDA_SENSORS
-    | ECOSOL_SENSORS
-    | ECOSOL_BINARY_SENSORS
-    | ECOSTER_BINARY_SENSORS
-    | set().union(*SENSOR_MIXER_KEY.values())
-    | MIXER_PUMP_BINARY_SENSOR_KEYS
-)
-
-# --- rmCurrentDataParams ID → regParams key mapping --------------------------
-# The rmCurrentDataParams endpoint provides only metadata (name, unit, special)
-# but NOT current values.  The actual live values are in regParams under
-# camelCase keys.  This mapping allows CustomSensor to read the value from
-# regParams when the user creates a custom entity from rmCurrentDataParams.
-CDP_ID_TO_REGPARAMS: dict[str, str] = {
-    # Monitoring – temperatures
-    "1024": "tempCO",
-    "1025": "tempCWU",
-    "1028": "tempUpperBuffer",
-    "1029": "tempLowerBuffer",
-    "1030": "tempFlueGas",
-    "1031": "mixerTemp1",
-    "1032": "mixerTemp2",
-    "1033": "mixerTemp3",
-    "1034": "mixerTemp4",
-    # Monitoring – low-ID sensors
-    "1": "lighter",
-    "26": "tempFeeder",
-    "28": "tempExternalSensor",
-    "114": "fuelLevel",
-    "117": "thermostat",
-    # Monitoring – pumps / actuators
-    "1536": "fan",
-    "1538": "feeder",
-    "1540": "feeder2Additional",
-    "1541": "pumpCO",
-    "1542": "pumpCWU",
-    "1543": "pumpCirculation",
-    "1544": "mixerPumpWorks1",
-    "1547": "mixerPumpWorks2",
-    "1550": "mixerPumpWorks3",
-    "1553": "mixerPumpWorks4",
-    # Monitoring – burner
-    "1794": "boilerPower",
-    # Setpoints (editable) – also in NUMBER_MAP but kept for completeness
-    "1280": "tempCOSet",
-    "1281": "tempCWUSet",
-    "1287": "mixerSetTemp1",
-    "1288": "mixerSetTemp2",
-    "1289": "mixerSetTemp3",
-    "1290": "mixerSetTemp4",
-}
-
-# IDs from rmCurrentDataParams that already have static entities.
-# Used to filter the Options Flow to prevent duplicate entity creation.
-STATIC_CDP_IDS: set[str] = set(CDP_ID_TO_REGPARAMS.keys()) & {
-    k for k, v in CDP_ID_TO_REGPARAMS.items() if v in STATIC_REGPARAMS_KEYS
-}
-
-# =============================================================================
-# CUSTOM ENTITY SELECTOR (Options Flow)
-# =============================================================================
-# Key used inside entry.options to store user-defined custom entities.
-CONF_CUSTOM_ENTITIES = "custom_entities"
-
-# Entity type constants for custom entity classification.
-CUSTOM_ENTITY_TYPE_SENSOR = "sensor"
-CUSTOM_ENTITY_TYPE_BINARY_SENSOR = "binary_sensor"
-
-# Component choices for the Options Flow dropdown, built from DEFAULT_COMPONENT_STATUS.
-CUSTOM_ENTITY_COMPONENTS: dict[str, str] = {
-    key: key.replace("_", " ").title() for key in DEFAULT_COMPONENT_STATUS
-}
-
-# Sensor-specific options for the Options Flow configure_sensor step.
-CUSTOM_SENSOR_UNIT_OPTIONS: dict[str | None, str] = {
-    None: "None (auto)",
-    **{unit: unit for unit in UNIT_NAME_TO_HA_UNIT},
-}
-
-CUSTOM_SENSOR_DEVICE_CLASS_OPTIONS: dict[str | None, str] = {
-    None: "None (auto)",
-    SensorDeviceClass.TEMPERATURE: "Temperature",
-    SensorDeviceClass.POWER: "Power",
-    SensorDeviceClass.POWER_FACTOR: "Power factor",
-    SensorDeviceClass.ENERGY: "Energy",
-    SensorDeviceClass.HUMIDITY: "Humidity",
-    SensorDeviceClass.PRESSURE: "Pressure",
-    SensorDeviceClass.DURATION: "Duration",
-    SensorDeviceClass.SIGNAL_STRENGTH: "Signal strength",
-    SensorDeviceClass.WEIGHT: "Weight",
-}
-
-CUSTOM_SENSOR_PRECISION_OPTIONS: dict[int | None, str] = {
-    None: "Auto",
-    0: "0 (integer)",
-    1: "0.0",
-    2: "0.00",
-    3: "0.000",
-}
-
-# =============================================================================
-# ENTITY UNIT MAPPINGS
-# =============================================================================
-# By default all sensors unit_of_measurement are None
-ENTITY_UNIT_MAP = {
-    "tempCO": UnitOfTemperature.CELSIUS,
-    "tempCOSet": UnitOfTemperature.CELSIUS,
-    "tempCWUSet": UnitOfTemperature.CELSIUS,
-    "tempExternalSensor": UnitOfTemperature.CELSIUS,
-    "tempFeeder": UnitOfTemperature.CELSIUS,
-    "lambdaLevel": PERCENTAGE,
-    "lambdaSet": PERCENTAGE,
-    "workAt100": UnitOfTime.HOURS,
-    "workAt50": UnitOfTime.HOURS,
-    "workAt30": UnitOfTime.HOURS,
-    "FeederWork": UnitOfTime.HOURS,
-    "thermoTemp": UnitOfTemperature.CELSIUS,
-    "fanPower": PERCENTAGE,
-    "tempFlueGas": UnitOfTemperature.CELSIUS,
-    "mixerSetTemp1": UnitOfTemperature.CELSIUS,
-    "mixerSetTemp2": UnitOfTemperature.CELSIUS,
-    "mixerSetTemp3": UnitOfTemperature.CELSIUS,
-    "mixerSetTemp4": UnitOfTemperature.CELSIUS,
-    "mixerSetTemp5": UnitOfTemperature.CELSIUS,
-    "mixerSetTemp6": UnitOfTemperature.CELSIUS,
-    "mixerTemp1": UnitOfTemperature.CELSIUS,
-    "mixerTemp2": UnitOfTemperature.CELSIUS,
-    "mixerTemp3": UnitOfTemperature.CELSIUS,
-    "mixerTemp4": UnitOfTemperature.CELSIUS,
-    "mixerTemp5": UnitOfTemperature.CELSIUS,
-    "mixerTemp6": UnitOfTemperature.CELSIUS,
-    "tempBack": UnitOfTemperature.CELSIUS,
-    "tempCWU": UnitOfTemperature.CELSIUS,
-    "boilerPower": PERCENTAGE,
-    "boilerPowerKW": UnitOfPower.KILO_WATT,
-    "fuelLevel": PERCENTAGE,
-    "tempUpperBuffer": UnitOfTemperature.CELSIUS,
-    "tempLowerBuffer": UnitOfTemperature.CELSIUS,
-    "signal": SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
-    "quality": PERCENTAGE,
-    "valveMixer1": PERCENTAGE,
-    "burnerOutput": PERCENTAGE,
-    "mixerTemp": UnitOfTemperature.CELSIUS,
-    "mixerSetTemp": UnitOfTemperature.CELSIUS,
-    # ecoMAX850R2-X specific units
-    "fuelConsum": PERCENTAGE,
-    SENSOR_FUEL_STREAM: "kg/h",  # Mass flow rate (confirmed from ecoNET cloud API)
-    "transmission": None,
-    # ecoSTER thermostat units
-    "ecoSterTemp1": UnitOfTemperature.CELSIUS,
-    "ecoSterTemp2": UnitOfTemperature.CELSIUS,
-    "ecoSterTemp3": UnitOfTemperature.CELSIUS,
-    "ecoSterTemp4": UnitOfTemperature.CELSIUS,
-    "ecoSterTemp5": UnitOfTemperature.CELSIUS,
-    "ecoSterTemp6": UnitOfTemperature.CELSIUS,
-    "ecoSterTemp7": UnitOfTemperature.CELSIUS,
-    "ecoSterTemp8": UnitOfTemperature.CELSIUS,
-    "ecoSterSetTemp1": UnitOfTemperature.CELSIUS,
-    "ecoSterSetTemp2": UnitOfTemperature.CELSIUS,
-    "ecoSterSetTemp3": UnitOfTemperature.CELSIUS,
-    "ecoSterSetTemp4": UnitOfTemperature.CELSIUS,
-    "ecoSterSetTemp5": UnitOfTemperature.CELSIUS,
-    "ecoSterSetTemp6": UnitOfTemperature.CELSIUS,
-    "ecoSterSetTemp7": UnitOfTemperature.CELSIUS,
-    "ecoSterSetTemp8": UnitOfTemperature.CELSIUS,
-    "ecoSterMode1": None,
-    "ecoSterMode2": None,
-    "ecoSterMode3": None,
-    "ecoSterMode4": None,
-    "ecoSterMode5": None,
-    "ecoSterMode6": None,
-    "ecoSterMode7": None,
-    "ecoSterMode8": None,
-    # ecoMAX360i
-    "Circuit2thermostatTemp": UnitOfTemperature.CELSIUS,
-    "Circuit3thermostatTemp": UnitOfTemperature.CELSIUS,
-    "Circuit4thermostatTemp": UnitOfTemperature.CELSIUS,
-    "Circuit5thermostatTemp": UnitOfTemperature.CELSIUS,
-    "Circuit6thermostatTemp": UnitOfTemperature.CELSIUS,
-    "Circuit7thermostatTemp": UnitOfTemperature.CELSIUS,
-    "TempClutch": UnitOfTemperature.CELSIUS,
-    "TempWthr": UnitOfTemperature.CELSIUS,
-    "TempCircuit2": UnitOfTemperature.CELSIUS,
-    "TempCircuit3": UnitOfTemperature.CELSIUS,
-    "TempCircuit4": UnitOfTemperature.CELSIUS,
-    "TempCircuit5": UnitOfTemperature.CELSIUS,
-    "TempCircuit6": UnitOfTemperature.CELSIUS,
-    "TempCircuit7": UnitOfTemperature.CELSIUS,
-    "TempBuforUp": UnitOfTemperature.CELSIUS,
-    "TempCWU": UnitOfTemperature.CELSIUS,
-    "TempBuforDown": UnitOfTemperature.CELSIUS,
-    "heatingUpperTemp": UnitOfTemperature.CELSIUS,
-    # Circuit1thermostat is a thermostat state (no "Temp" suffix), not a temperature.
-    "Circuit1thermostat": None,
-    "flapValveStates": None,
-    "HeatDemanded": None,
-    "WaterPumpRunning": None,
-    # Heat pump (Axen) units
-    "AxenOutdoorTemp": UnitOfTemperature.CELSIUS,
-    "AxenOutgoingTemp": UnitOfTemperature.CELSIUS,
-    "AxenReturnTemp": UnitOfTemperature.CELSIUS,
-    "AxenCompressorFreq": "Hz",
-    "AxenUpperPump": None,
-    "AxenWorkState": None,
-    "HeatSourceCalcPresetTemp": UnitOfTemperature.CELSIUS,
-    # Heat pump compressor temperature units
-    "afterCompressorTemp": UnitOfTemperature.CELSIUS,
-    "beforeCompressorTemp": UnitOfTemperature.CELSIUS,
-    "exhaustGasTemp": UnitOfTemperature.CELSIUS,
-    "outdoorTemp": UnitOfTemperature.CELSIUS,
-    # Heat pump status units
-    "HPStatusWorkMode": None,
-    "HPStatusPresetTemp": UnitOfTemperature.CELSIUS,
-    "HPStatusControl": None,
-    # SSA (weather compensation) units
-    "ssaState": None,
-    "ssaCorr": UnitOfTemperature.CELSIUS,
-    "ssaPrevTemp": UnitOfTemperature.CELSIUS,
-    # Circuit comfort/eco temperature setpoints
-    "Circuit1ComfortTemp": UnitOfTemperature.CELSIUS,
-    "Circuit1EcoTemp": UnitOfTemperature.CELSIUS,
-    "BufferTargetTemp": UnitOfTemperature.CELSIUS,
-    "Circuit2ComfortTemp": UnitOfTemperature.CELSIUS,
-    "Circuit2EcoTemp": UnitOfTemperature.CELSIUS,
-    "Circuit3ComfortTemp": UnitOfTemperature.CELSIUS,
-    "Circuit3EcoTemp": UnitOfTemperature.CELSIUS,
-    "Circuit4ComfortTemp": UnitOfTemperature.CELSIUS,
-    "Circuit4EcoTemp": UnitOfTemperature.CELSIUS,
-    "Circuit5ComfortTemp": UnitOfTemperature.CELSIUS,
-    "Circuit5EcoTemp": UnitOfTemperature.CELSIUS,
-    "Circuit6ComfortTemp": UnitOfTemperature.CELSIUS,
-    "Circuit6EcoTemp": UnitOfTemperature.CELSIUS,
-    "Circuit7ComfortTemp": UnitOfTemperature.CELSIUS,
-    "Circuit7EcoTemp": UnitOfTemperature.CELSIUS,
-    # ecoMAX360i informationParams / editParams sensors
-    "TargetFlowTemp": UnitOfTemperature.CELSIUS,
-    "ActualFlowTemp": UnitOfTemperature.CELSIUS,
-    "ActualReturnTemp": UnitOfTemperature.CELSIUS,
-    "FanSpeed": "rpm",
-    "HeatPumpAmbient": UnitOfTemperature.CELSIUS,
-    "ActualDHWTemp": UnitOfTemperature.CELSIUS,
-    "Circuit1DesiredLWT": UnitOfTemperature.CELSIUS,
-    "ElectricalPower": UnitOfPower.KILO_WATT,
-    "ThermalPower": UnitOfPower.KILO_WATT,
-    "COP": None,
-    "SCOP": None,
-    "FlowRate": "L/min",
-    "AXENREGISTER64": None,
-    "AXENREGISTER65": None,
-    # Phoenix / extended heat-pump units (decomposed from EXTRA_SENSORS)
-    "BuforCalcSetTemp": UnitOfTemperature.CELSIUS,
-    "Circuit1CalcTemp": UnitOfTemperature.CELSIUS,
-    "Circuit2CalcTemp": UnitOfTemperature.CELSIUS,
-    "Circuit4CalcTemp": UnitOfTemperature.CELSIUS,
-    "Circuit5CalcTemp": UnitOfTemperature.CELSIUS,
-    "Circuit6CalcTemp": UnitOfTemperature.CELSIUS,
-    "Circuit7CalcTemp": UnitOfTemperature.CELSIUS,
-    "HDWsetpointcalculate": UnitOfTemperature.CELSIUS,
-    "CoolingPower": UnitOfPower.KILO_WATT,
-    "HeatingPower": UnitOfPower.KILO_WATT,
-    "ElectricPower": UnitOfPower.KILO_WATT,
-    "currentFlow": "m³/h",
-    "PHNXcoil_temp": UnitOfTemperature.CELSIUS,
-    "PHNXdischarge_temp": UnitOfTemperature.CELSIUS,
-    "PHNXinletTemp": UnitOfTemperature.CELSIUS,
-    "PHNXinletTemp_pump": UnitOfTemperature.CELSIUS,
-    "PHNXinletTemp_pump2": UnitOfTemperature.CELSIUS,
-    "PHNXinletTemp_pump3": UnitOfTemperature.CELSIUS,
-    "PHNXinletTemp_pump5": UnitOfTemperature.CELSIUS,
-    "PHNXoutletTemp": UnitOfTemperature.CELSIUS,
-    "PHNXoutletTemp_pump2": UnitOfTemperature.CELSIUS,
-    "PHNXoutletTemp_pump3": UnitOfTemperature.CELSIUS,
-    "PHNXoutletTemp_pump4": UnitOfTemperature.CELSIUS,
-    "PHNXoutletTemp_pump5": UnitOfTemperature.CELSIUS,
-    "PHNXreg2045": UnitOfTemperature.CELSIUS,
-    "PHNXreg2046": UnitOfTemperature.CELSIUS,
-    "PHNXreg2071": "Hz",
-    "PHNXreg2074": "rpm",
-    "PHNXsuctionTemp": UnitOfTemperature.CELSIUS,
-    # ecoSOL specific units (ecoSOL [n] models)
-    "T1": UnitOfTemperature.CELSIUS,
-    "T2": UnitOfTemperature.CELSIUS,
-    "T3": UnitOfTemperature.CELSIUS,
-    "T4": UnitOfTemperature.CELSIUS,
-    "T5": UnitOfTemperature.CELSIUS,
-    "T6": UnitOfTemperature.CELSIUS,
-    "TzCWU": UnitOfTemperature.CELSIUS,
-    "P1": None,
-    "P2": None,
-    "H": None,
-    "Moc_chwilowa": UnitOfPower.KILO_WATT,
-    "Uzysk_ca_kowity": UnitOfEnergy.KILO_WATT_HOUR,
-    # ecoSOL diagnostic units
-    "ecosrvAddr": None,
-    "softVer": None,
-    "routerType": None,
-    "protocolType": None,
-    "controllerID": None,
-    "ecosrvSoftVer": None,
-    "moduleEcoSTERSoftVer": None,
-}
-
-# =============================================================================
-# ENTITY STATE CLASS MAPPINGS
-# =============================================================================
-# By default all sensors state_class are MEASUREMENT
-STATE_CLASS_MAP: dict[str, SensorStateClass | None] = {
-    "lambdaStatus": None,
-    "mode": None,
-    "thermostat": None,
-    "statusCO": None,
-    "statusCWU": None,
-    "softVer": None,
-    "controllerID": None,
-    "moduleASoftVer": None,
-    "moduleBSoftVer": None,
-    "moduleCSoftVer": None,
-    "moduleLambdaSoftVer": None,
-    "modulePanelSoftVer": None,
-    # Diagnostic sensors (non-numeric)
-    "routerType": None,
-    "protocolType": None,
-    "moduleEcoSTERSoftVer": None,
-    "ecosrvSoftVer": None,
-    "transmission": None,
-    "ecoSterMode1": None,
-    "ecoSterMode2": None,
-    "ecoSterMode3": None,
-    "ecoSterMode4": None,
-    "ecoSterMode5": None,
-    "ecoSterMode6": None,
-    "ecoSterMode7": None,
-    "ecoSterMode8": None,
-    "P1": None,
-    "P2": None,
-    "H": None,
-    "Moc_chwilowa": SensorStateClass.MEASUREMENT,
-    "Uzysk_ca_kowity": SensorStateClass.TOTAL_INCREASING,
-    # ecoMAX360i
-    "PS": None,
-    "heating_work_state_pump4": None,
-    "flapValveStates": None,
-    "HeatDemanded": None,
-    "WaterPumpRunning": None,
-    # ecoMAX360i informationParams / editParams sensors
-    "FanSpeed": None,
-    "COP": None,
-    "SCOP": None,
-    "FlowRate": None,
-    "AXENREGISTER64": None,
-    "AXENREGISTER65": None,
-    # ecoMAX360i heat pump status sensors (not measurements)
-    "HPStatusWorkMode": None,
-    "HPStatusControl": None,
-    "AxenUpperPump": None,
-    "AxenWorkState": None,
-    "ssaState": None,
-    # Phoenix / extended heat-pump status sensors (not measurements)
-    "Circuit1thermostat": None,
-    "HPStatusBuffHeatStat": None,
-    "HPStatusCircPStat0": None,
-    "HPStatusCircPStat1": None,
-    "HPStatusCircPStat2": None,
-    "HPStatusCircPStat3": None,
-    "HPStatusCircPStat4": None,
-    "HPStatusCircPStat5": None,
-    "HPStatusCircPStat6": None,
-    "HPStatusFlowHeatStat": None,
-    "HPStatusHdwHeatStat": None,
-    "HPStatusUhsStat": None,
-    "HP_work_state_set_pump2": None,
-    "HP_work_state_set_pump3": None,
-    "HP_work_state_set_pump4": None,
-    "HP_work_state_set_pump5": None,
-    # Phoenix / extended heat-pump measurements (numeric; explicit because the
-    # default state_class is None to avoid bogus statistics on all-sensors keys).
-    "BuforCalcSetTemp": SensorStateClass.MEASUREMENT,
-    "Circuit1CalcTemp": SensorStateClass.MEASUREMENT,
-    "Circuit2CalcTemp": SensorStateClass.MEASUREMENT,
-    "Circuit4CalcTemp": SensorStateClass.MEASUREMENT,
-    "Circuit5CalcTemp": SensorStateClass.MEASUREMENT,
-    "Circuit6CalcTemp": SensorStateClass.MEASUREMENT,
-    "Circuit7CalcTemp": SensorStateClass.MEASUREMENT,
-    "Circuit2thermostatTemp": SensorStateClass.MEASUREMENT,
-    "Circuit3thermostatTemp": SensorStateClass.MEASUREMENT,
-    "HDWsetpointcalculate": SensorStateClass.MEASUREMENT,
-    "HPStatusPresetTemp": SensorStateClass.MEASUREMENT,
-    "CoolingPower": SensorStateClass.MEASUREMENT,
-    "HeatingPower": SensorStateClass.MEASUREMENT,
-    "ElectricPower": SensorStateClass.MEASUREMENT,
-    "currentFlow": SensorStateClass.MEASUREMENT,
-    "TempBuforDown": SensorStateClass.MEASUREMENT,
-    "TempBuforUp": SensorStateClass.MEASUREMENT,
-    "TempCWU": SensorStateClass.MEASUREMENT,
-    "TempCircuit2": SensorStateClass.MEASUREMENT,
-    "TempCircuit3": SensorStateClass.MEASUREMENT,
-    "TempCircuit4": SensorStateClass.MEASUREMENT,
-    "TempCircuit5": SensorStateClass.MEASUREMENT,
-    "TempCircuit6": SensorStateClass.MEASUREMENT,
-    "TempCircuit7": SensorStateClass.MEASUREMENT,
-    "TempClutch": SensorStateClass.MEASUREMENT,
-    "TempWthr": SensorStateClass.MEASUREMENT,
-    "PHNXcoil_temp": SensorStateClass.MEASUREMENT,
-    "PHNXdischarge_temp": SensorStateClass.MEASUREMENT,
-    "PHNXinletTemp": SensorStateClass.MEASUREMENT,
-    "PHNXinletTemp_pump": SensorStateClass.MEASUREMENT,
-    "PHNXinletTemp_pump2": SensorStateClass.MEASUREMENT,
-    "PHNXinletTemp_pump3": SensorStateClass.MEASUREMENT,
-    "PHNXinletTemp_pump5": SensorStateClass.MEASUREMENT,
-    "PHNXoutletTemp": SensorStateClass.MEASUREMENT,
-    "PHNXoutletTemp_pump2": SensorStateClass.MEASUREMENT,
-    "PHNXoutletTemp_pump3": SensorStateClass.MEASUREMENT,
-    "PHNXoutletTemp_pump4": SensorStateClass.MEASUREMENT,
-    "PHNXoutletTemp_pump5": SensorStateClass.MEASUREMENT,
-    "PHNXreg2045": SensorStateClass.MEASUREMENT,
-    "PHNXreg2046": SensorStateClass.MEASUREMENT,
-    "PHNXreg2071": SensorStateClass.MEASUREMENT,
-    "PHNXreg2074": SensorStateClass.MEASUREMENT,
-    "PHNXsuctionTemp": SensorStateClass.MEASUREMENT,
-}
-
-# =============================================================================
-# ENTITY DEVICE CLASS MAPPINGS
-# =============================================================================
-
-ENTITY_SENSOR_DEVICE_CLASS_MAP: dict[str, SensorDeviceClass | None] = {
-    "tempFeeder": SensorDeviceClass.TEMPERATURE,
-    "tempExternalSensor": SensorDeviceClass.TEMPERATURE,
-    "tempCO": SensorDeviceClass.TEMPERATURE,
-    "tempCOSet": SensorDeviceClass.TEMPERATURE,
-    "boilerPower": SensorDeviceClass.POWER_FACTOR,
-    "boilerPowerKW": SensorDeviceClass.POWER,
-    "fanPower": SensorDeviceClass.POWER_FACTOR,
-    "tempFlueGas": SensorDeviceClass.TEMPERATURE,
-    "mixerSetTemp1": SensorDeviceClass.TEMPERATURE,
-    "mixerSetTemp2": SensorDeviceClass.TEMPERATURE,
-    "mixerSetTemp3": SensorDeviceClass.TEMPERATURE,
-    "mixerSetTemp4": SensorDeviceClass.TEMPERATURE,
-    "mixerSetTemp5": SensorDeviceClass.TEMPERATURE,
-    "mixerSetTemp6": SensorDeviceClass.TEMPERATURE,
-    "mixerTemp1": SensorDeviceClass.TEMPERATURE,
-    "mixerTemp2": SensorDeviceClass.TEMPERATURE,
-    "mixerTemp3": SensorDeviceClass.TEMPERATURE,
-    "mixerTemp4": SensorDeviceClass.TEMPERATURE,
-    "mixerTemp5": SensorDeviceClass.TEMPERATURE,
-    "mixerTemp6": SensorDeviceClass.TEMPERATURE,
-    "mixerTemp": SensorDeviceClass.TEMPERATURE,
-    "mixerSetTemp": SensorDeviceClass.TEMPERATURE,
-    "tempBack": SensorDeviceClass.TEMPERATURE,
-    "tempCWU": SensorDeviceClass.TEMPERATURE,
-    "statusCO": SensorDeviceClass.ENUM,
-    "statusCWU": None,
-    "tempUpperBuffer": SensorDeviceClass.TEMPERATURE,
-    "tempLowerBuffer": SensorDeviceClass.TEMPERATURE,
-    "signal": SensorDeviceClass.SIGNAL_STRENGTH,
-    "servoMixer1": SensorDeviceClass.ENUM,
-    "mode": SensorDeviceClass.ENUM,
-    "transmission": SensorDeviceClass.ENUM,
-    # ecoMAX850R2-X specific device classes
-    "fuelConsum": SensorDeviceClass.POWER_FACTOR,
-    # fuelStream: No device class - HA has no SensorDeviceClass for mass flow rate (kg/h)
-    # ecoSTER thermostat device classes
-    "ecoSterTemp1": SensorDeviceClass.TEMPERATURE,
-    "ecoSterTemp2": SensorDeviceClass.TEMPERATURE,
-    "ecoSterTemp3": SensorDeviceClass.TEMPERATURE,
-    "ecoSterTemp4": SensorDeviceClass.TEMPERATURE,
-    "ecoSterTemp5": SensorDeviceClass.TEMPERATURE,
-    "ecoSterTemp6": SensorDeviceClass.TEMPERATURE,
-    "ecoSterTemp7": SensorDeviceClass.TEMPERATURE,
-    "ecoSterTemp8": SensorDeviceClass.TEMPERATURE,
-    "ecoSterSetTemp1": SensorDeviceClass.TEMPERATURE,
-    "ecoSterSetTemp2": SensorDeviceClass.TEMPERATURE,
-    "ecoSterSetTemp3": SensorDeviceClass.TEMPERATURE,
-    "ecoSterSetTemp4": SensorDeviceClass.TEMPERATURE,
-    "ecoSterSetTemp5": SensorDeviceClass.TEMPERATURE,
-    "ecoSterSetTemp6": SensorDeviceClass.TEMPERATURE,
-    "ecoSterSetTemp7": SensorDeviceClass.TEMPERATURE,
-    "ecoSterSetTemp8": SensorDeviceClass.TEMPERATURE,
-    "ecoSterMode1": None,
-    "ecoSterMode2": None,
-    "ecoSterMode3": None,
-    "ecoSterMode4": None,
-    "ecoSterMode5": None,
-    "ecoSterMode6": None,
-    "ecoSterMode7": None,
-    "ecoSterMode8": None,
-    # ecoMAX360i
-    "Circuit2thermostatTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit3thermostatTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit4thermostatTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit5thermostatTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit6thermostatTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit7thermostatTemp": SensorDeviceClass.TEMPERATURE,
-    "TempClutch": SensorDeviceClass.TEMPERATURE,
-    "TempWthr": SensorDeviceClass.TEMPERATURE,
-    "TempCircuit2": SensorDeviceClass.TEMPERATURE,
-    "TempCircuit3": SensorDeviceClass.TEMPERATURE,
-    "TempCircuit4": SensorDeviceClass.TEMPERATURE,
-    "TempCircuit5": SensorDeviceClass.TEMPERATURE,
-    "TempCircuit6": SensorDeviceClass.TEMPERATURE,
-    "TempCircuit7": SensorDeviceClass.TEMPERATURE,
-    "TempBuforUp": SensorDeviceClass.TEMPERATURE,
-    "TempCWU": SensorDeviceClass.TEMPERATURE,
-    "TempBuforDown": SensorDeviceClass.TEMPERATURE,
-    "heatingUpperTemp": SensorDeviceClass.TEMPERATURE,
-    # Circuit1thermostat is a thermostat state (no "Temp" suffix), not a temperature.
-    "Circuit1thermostat": None,
-    "flapValveStates": SensorDeviceClass.ENUM,
-    "HeatDemanded": SensorDeviceClass.ENUM,
-    "WaterPumpRunning": SensorDeviceClass.ENUM,
-    # Heat pump (Axen) device classes
-    "AxenOutdoorTemp": SensorDeviceClass.TEMPERATURE,
-    "AxenOutgoingTemp": SensorDeviceClass.TEMPERATURE,
-    "AxenReturnTemp": SensorDeviceClass.TEMPERATURE,
-    "AxenCompressorFreq": SensorDeviceClass.FREQUENCY,
-    "AxenUpperPump": None,
-    "AxenWorkState": None,
-    "HeatSourceCalcPresetTemp": SensorDeviceClass.TEMPERATURE,
-    # Heat pump compressor temperature device classes
-    "afterCompressorTemp": SensorDeviceClass.TEMPERATURE,
-    "beforeCompressorTemp": SensorDeviceClass.TEMPERATURE,
-    "exhaustGasTemp": SensorDeviceClass.TEMPERATURE,
-    "outdoorTemp": SensorDeviceClass.TEMPERATURE,
-    # Heat pump status device classes
-    "HPStatusWorkMode": None,
-    "HPStatusPresetTemp": SensorDeviceClass.TEMPERATURE,
-    "HPStatusControl": None,
-    # SSA (weather compensation) device classes
-    "ssaState": None,
-    "ssaCorr": SensorDeviceClass.TEMPERATURE,
-    "ssaPrevTemp": SensorDeviceClass.TEMPERATURE,
-    # Circuit comfort/eco temperature device classes
-    "Circuit1ComfortTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit1EcoTemp": SensorDeviceClass.TEMPERATURE,
-    "BufferTargetTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit2ComfortTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit2EcoTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit3ComfortTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit3EcoTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit4ComfortTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit4EcoTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit5ComfortTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit5EcoTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit6ComfortTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit6EcoTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit7ComfortTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit7EcoTemp": SensorDeviceClass.TEMPERATURE,
-    # ecoMAX360i informationParams / editParams device classes
-    "TargetFlowTemp": SensorDeviceClass.TEMPERATURE,
-    "ActualFlowTemp": SensorDeviceClass.TEMPERATURE,
-    "ActualReturnTemp": SensorDeviceClass.TEMPERATURE,
-    "FanSpeed": None,
-    "HeatPumpAmbient": SensorDeviceClass.TEMPERATURE,
-    "ActualDHWTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit1DesiredLWT": SensorDeviceClass.TEMPERATURE,
-    "ElectricalPower": SensorDeviceClass.POWER,
-    "ThermalPower": SensorDeviceClass.POWER,
-    "COP": SensorDeviceClass.POWER_FACTOR,
-    "SCOP": SensorDeviceClass.POWER_FACTOR,
-    "FlowRate": None,
-    "AXENREGISTER64": None,
-    "AXENREGISTER65": None,
-    # Phoenix / extended heat-pump device classes (decomposed from EXTRA_SENSORS)
-    "BuforCalcSetTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit1CalcTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit2CalcTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit4CalcTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit5CalcTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit6CalcTemp": SensorDeviceClass.TEMPERATURE,
-    "Circuit7CalcTemp": SensorDeviceClass.TEMPERATURE,
-    "HDWsetpointcalculate": SensorDeviceClass.TEMPERATURE,
-    "CoolingPower": SensorDeviceClass.POWER,
-    "HeatingPower": SensorDeviceClass.POWER,
-    "ElectricPower": SensorDeviceClass.POWER,
-    "PHNXcoil_temp": SensorDeviceClass.TEMPERATURE,
-    "PHNXdischarge_temp": SensorDeviceClass.TEMPERATURE,
-    "PHNXinletTemp": SensorDeviceClass.TEMPERATURE,
-    "PHNXinletTemp_pump": SensorDeviceClass.TEMPERATURE,
-    "PHNXinletTemp_pump2": SensorDeviceClass.TEMPERATURE,
-    "PHNXinletTemp_pump3": SensorDeviceClass.TEMPERATURE,
-    "PHNXinletTemp_pump5": SensorDeviceClass.TEMPERATURE,
-    "PHNXoutletTemp": SensorDeviceClass.TEMPERATURE,
-    "PHNXoutletTemp_pump2": SensorDeviceClass.TEMPERATURE,
-    "PHNXoutletTemp_pump3": SensorDeviceClass.TEMPERATURE,
-    "PHNXoutletTemp_pump4": SensorDeviceClass.TEMPERATURE,
-    "PHNXoutletTemp_pump5": SensorDeviceClass.TEMPERATURE,
-    "PHNXreg2045": SensorDeviceClass.TEMPERATURE,
-    "PHNXreg2046": SensorDeviceClass.TEMPERATURE,
-    "PHNXreg2071": SensorDeviceClass.FREQUENCY,
-    "PHNXsuctionTemp": SensorDeviceClass.TEMPERATURE,
-    # ecoSOL specific device classes (ecoSOL [n] models)
-    "T1": SensorDeviceClass.TEMPERATURE,
-    "T2": SensorDeviceClass.TEMPERATURE,
-    "T3": SensorDeviceClass.TEMPERATURE,
-    "T4": SensorDeviceClass.TEMPERATURE,
-    "T5": SensorDeviceClass.TEMPERATURE,
-    "T6": SensorDeviceClass.TEMPERATURE,
-    "TzCWU": SensorDeviceClass.TEMPERATURE,
-    "P1": None,
-    "P2": None,
-    "H": None,
-    "Moc_chwilowa": SensorDeviceClass.POWER,
-    "Uzysk_ca_kowity": SensorDeviceClass.ENERGY,
-    # ecoSOL diagnostic device classes
-    "ecosrvAddr": None,
-    "softVer": None,
-    "routerType": None,
-    "protocolType": None,
-    "controllerID": None,
-    "ecosrvSoftVer": None,
-}
-
-# Number entity device classes
-ENTITY_NUMBER_SENSOR_DEVICE_CLASS_MAP = {
-    "tempCOSet": NumberDeviceClass.TEMPERATURE,
-    "tempCWUSet": NumberDeviceClass.TEMPERATURE,
-    # ecoMAX850R2-X specific number entities
-    "mixerSetTemp5": NumberDeviceClass.TEMPERATURE,
-}
-
-# Binary sensor device classes
-ENTITY_BINARY_DEVICE_CLASS_MAP = {
-    # By default all binary sensors device_class are RUNNING
-    "mainSrv": BinarySensorDeviceClass.CONNECTIVITY,
-    "wifi": BinarySensorDeviceClass.CONNECTIVITY,
-    "lan": BinarySensorDeviceClass.CONNECTIVITY,
-    "fuelConsumptionCalc": BinarySensorDeviceClass.RUNNING,
-    "ecosrvHttps": BinarySensorDeviceClass.CONNECTIVITY,
-    # ecoMAX850R2-X specific binary sensors
-    "contactGZC": BinarySensorDeviceClass.CONNECTIVITY,
-    "contactGZCActive": BinarySensorDeviceClass.CONNECTIVITY,
-    # Alarm sensors
-    "alarmActive": BinarySensorDeviceClass.PROBLEM,
-}
-
-# =============================================================================
-# ENTITY PRECISION MAPPINGS
-# =============================================================================
-# Add only keys where precision more than 0 needed
-ENTITY_PRECISION = {
-    "tempFeeder": 1,
-    "tempExternalSensor": 1,
-    "lambdaLevel": 1,
-    "lambdaSet": 1,
-    "tempCO": 1,
-    "mixerTemp1": 1,
-    "mixerTemp2": 1,
-    "mixerTemp3": 1,
-    "mixerTemp4": 1,
-    "mixerTemp5": 1,
-    "mixerTemp6": 1,
-    "tempBack": 2,
-    "tempUpperBuffer": 1,
-    "tempLowerBuffer": 1,
-    "tempCWU": 1,
-    "tempFlueGas": 1,
-    "fanPower": 0,
-    "statusCO": None,
-    "statusCWU": None,
-    "thermostat": None,
-    "lambdaStatus": None,
-    "mode": None,
-    "transmission": None,
-    "softVer": None,
-    "controllerID": None,
-    "moduleASoftVer": None,
-    "moduleBSoftVer": None,
-    "moduleCSoftVer": None,
-    "moduleLambdaSoftVer": None,
-    "modulePanelSoftVer": None,
-    "moduleEcoSTERSoftVer": None,
-    "ecosrvSoftVer": None,
-    # ecoMAX360i
-    "PS": None,
-    "Circuit2thermostatTemp": 1,
-    "Circuit3thermostatTemp": 1,
-    "Circuit4thermostatTemp": 1,
-    "Circuit5thermostatTemp": 1,
-    "Circuit6thermostatTemp": 1,
-    "Circuit7thermostatTemp": 1,
-    "TempClutch": 1,
-    "TempWthr": 1,
-    "TempCircuit2": 1,
-    "TempCircuit3": 1,
-    "TempCircuit4": 1,
-    "TempCircuit5": 1,
-    "TempCircuit6": 1,
-    "TempCircuit7": 1,
-    "TempBuforUp": 1,
-    "TempCWU": 1,
-    "TempBuforDown": 1,
-    "flapValveStates": None,
-    "HeatDemanded": None,
-    "WaterPumpRunning": None,
-    # Heat pump (Axen) precision
-    "AxenOutdoorTemp": 1,
-    "AxenOutgoingTemp": 1,
-    "AxenReturnTemp": 1,
-    "AxenCompressorFreq": 0,
-    # Circuit comfort/eco temperature precision
-    "Circuit1ComfortTemp": 1,
-    "Circuit1EcoTemp": 1,
-    "BufferTargetTemp": 1,
-    "Circuit2ComfortTemp": 1,
-    "Circuit2EcoTemp": 1,
-    "Circuit3ComfortTemp": 1,
-    "Circuit3EcoTemp": 1,
-    "Circuit4ComfortTemp": 1,
-    "Circuit4EcoTemp": 1,
-    "Circuit5ComfortTemp": 1,
-    "Circuit5EcoTemp": 1,
-    "Circuit6ComfortTemp": 1,
-    "Circuit6EcoTemp": 1,
-    "Circuit7ComfortTemp": 1,
-    "Circuit7EcoTemp": 1,
-    # ecoMAX360i informationParams / editParams precision
-    "TargetFlowTemp": 1,
-    "ActualFlowTemp": 1,
-    "ActualReturnTemp": 1,
-    "FanSpeed": 0,
-    "HeatPumpAmbient": 1,
-    "ActualDHWTemp": 1,
-    "Circuit1DesiredLWT": 1,
-    "ElectricalPower": 2,
-    "ThermalPower": 2,
-    "COP": 2,
-    "SCOP": 2,
-    "FlowRate": 1,
-    "AXENREGISTER64": 1,
-    "AXENREGISTER65": 0,
-    # Heat pump compressor temperature precision
-    "afterCompressorTemp": 1,
-    "beforeCompressorTemp": 1,
-    "exhaustGasTemp": 1,
-    "outdoorTemp": 1,
-    # Heat pump status precision
-    "HPStatusWorkMode": None,
-    "HPStatusPresetTemp": 1,
-    "HPStatusControl": None,
-    "AxenUpperPump": None,
-    "AxenWorkState": None,
-    "HeatSourceCalcPresetTemp": 0,
-    # SSA (weather compensation) precision
-    "ssaState": None,
-    "ssaCorr": 0,
-    "ssaPrevTemp": 1,
-    # ecoSTER thermostat precision
-    "ecoSterTemp1": 1,
-    "ecoSterTemp2": 1,
-    "ecoSterTemp3": 1,
-    "ecoSterTemp4": 1,
-    "ecoSterTemp5": 1,
-    "ecoSterTemp6": 1,
-    "ecoSterTemp7": 1,
-    "ecoSterTemp8": 1,
-    "ecoSterSetTemp1": 1,
-    "ecoSterSetTemp2": 1,
-    "ecoSterSetTemp3": 1,
-    "ecoSterSetTemp4": 1,
-    "ecoSterSetTemp5": 1,
-    "ecoSterSetTemp6": 1,
-    "ecoSterSetTemp7": 1,
-    "ecoSterSetTemp8": 1,
-    "heatingUpperTemp": 1,
-    # Circuit1thermostat is a thermostat state (no "Temp" suffix), not a temperature.
-    "Circuit1thermostat": None,
-    "heating_work_state_pump4": None,
-    # ecoSOL specific precision (ecoSOL [n] models)
-    "T1": 1,
-    "T2": 1,
-    "T3": 1,
-    "T4": 1,
-    "T5": 1,
-    "T6": 1,
-    "TzCWU": 1,
-    "P1": None,
-    "P2": None,
-    "H": None,
-    "Moc_chwilowa": 2,
-    "Uzysk_ca_kowity": 2,
-    # ecoSOL diagnostic precision
-    "ecosrvAddr": None,
-    "routerType": None,
-    "protocolType": None,
-    # Phoenix / extended heat-pump precision (decomposed from EXTRA_SENSORS)
-    "BuforCalcSetTemp": 1,
-    "Circuit1CalcTemp": 1,
-    "Circuit2CalcTemp": 1,
-    "Circuit4CalcTemp": 1,
-    "Circuit5CalcTemp": 1,
-    "Circuit6CalcTemp": 1,
-    "Circuit7CalcTemp": 1,
-    "HDWsetpointcalculate": 0,
-    "CoolingPower": 2,
-    "HeatingPower": 2,
-    "ElectricPower": 2,
-    "currentFlow": 2,
-    "PHNXcoil_temp": 1,
-    "PHNXdischarge_temp": 1,
-    "PHNXinletTemp": 1,
-    "PHNXinletTemp_pump": 1,
-    "PHNXinletTemp_pump2": 1,
-    "PHNXinletTemp_pump3": 1,
-    "PHNXinletTemp_pump5": 1,
-    "PHNXoutletTemp": 1,
-    "PHNXoutletTemp_pump2": 1,
-    "PHNXoutletTemp_pump3": 1,
-    "PHNXoutletTemp_pump4": 1,
-    "PHNXoutletTemp_pump5": 1,
-    "PHNXreg2045": 1,
-    "PHNXreg2046": 1,
-    "PHNXreg2071": 1,
-    "PHNXreg2074": 0,
-    "PHNXsuctionTemp": 1,
-}
-
-NO_CWU_TEMP_SET_STATUS_CODE = 128
-
-ECOMAX360I_NUMERIC_SENSOR_PROCESSOR_KEYS: Final[tuple[str, ...]] = (
-    "ActualDHWTemp",
-    "ActualFlowTemp",
-    "ActualReturnTemp",
-    "AXENREGISTER64",
-    "AXENREGISTER65",
-    "Circuit1DesiredLWT",
-    "COP",
-    "ElectricalPower",
-    "FanSpeed",
-    "FlowRate",
-    "HeatPumpAmbient",
-    "HeatSourceCalcPresetTemp",
-    "HPStatusPresetTemp",
-    "SCOP",
-    "TargetFlowTemp",
-    "ThermalPower",
-    "afterCompressorTemp",
-    "beforeCompressorTemp",
-    "exhaustGasTemp",
-    "outdoorTemp",
-    "ssaCorr",
-    "ssaPrevTemp",
-)
-
-
-def _int_enum_lookup(mapping: dict[int, str], value: Any) -> str:
-    """Look up an integer-keyed enum mapping, coercing string values from informationParams."""
-    try:
-        return mapping.get(int(value), STATE_UNKNOWN)
-    except (TypeError, ValueError):
-        return STATE_UNKNOWN
-
-
-# scale-phnxreg2045-2046
-def _numeric_div10_or_none(value: Any) -> float | None:
-    """Return numeric value divided by 10, or None when unavailable."""
-    numeric = _numeric_or_none(value)
-    return None if numeric is None else numeric / 10
-
-
-def _numeric_or_none(value: Any) -> float | None:
-    """Return a numeric sensor value, or None when the controller reports a state."""
-    if isinstance(value, bool) or value is None:
-        return None
-    try:
-        return float(value)
-    except (TypeError, ValueError):
-        return None
-
-
-ENTITY_VALUE_PROCESSOR = {
-    **dict.fromkeys(ECOMAX360I_NUMERIC_SENSOR_PROCESSOR_KEYS, _numeric_or_none),
-    "mode": lambda x: SENSOR_MODE_MAPPING.get(x, STATE_UNKNOWN),
-    "lambdaStatus": lambda x: SENSOR_LAMBDA_STATUS_MAPPING.get(x, STATE_UNKNOWN),
-    "statusCWU": lambda x: SENSOR_STATUS_CWU_MAPPING.get(x, STATE_UNKNOWN),
-    "statusCO": lambda x: SENSOR_STATUS_CO_MAPPING.get(x, STATE_UNKNOWN),
-    "thermostat": lambda x: SENSOR_THERMOSTAT_MAPPING.get(x, STATE_UNKNOWN),
-    "transmission": lambda x: OPERATION_MODE_NAMES.get(x, STATE_UNKNOWN),
-    "PHNXreg2045": _numeric_div10_or_none,
-    "PHNXreg2046": _numeric_div10_or_none,
-    # ecoMAX360i-specific processors (informationParams yields string values)
-    "flapValveStates": lambda x: _int_enum_lookup(SENSOR_FLAP_VALVE_STATES_MAPPING, x),
-    "HeatDemanded": lambda x: _int_enum_lookup(SENSOR_HEAT_DEMANDED_MAPPING, x),
-    "WaterPumpRunning": lambda x: _int_enum_lookup(
-        SENSOR_WATER_PUMP_RUNNING_MAPPING, x
-    ),
-}
-
-# =============================================================================
-# ENTITY CATEGORY MAPPINGS
-# =============================================================================
-ENTITY_CATEGORY = {
-    "signal": EntityCategory.DIAGNOSTIC,
-    "quality": EntityCategory.DIAGNOSTIC,
-    "softVer": EntityCategory.DIAGNOSTIC,
-    "moduleASoftVer": EntityCategory.DIAGNOSTIC,
-    "moduleBSoftVer": EntityCategory.DIAGNOSTIC,
-    "modulePanelSoftVer": EntityCategory.DIAGNOSTIC,
-    "moduleLambdaSoftVer": EntityCategory.DIAGNOSTIC,
-    "protocolType": EntityCategory.DIAGNOSTIC,
-    "controllerID": EntityCategory.DIAGNOSTIC,
-    "moduleCSoftVer": EntityCategory.DIAGNOSTIC,
-    "mainSrv": EntityCategory.DIAGNOSTIC,
-    "wifi": EntityCategory.DIAGNOSTIC,
-    "lan": EntityCategory.DIAGNOSTIC,
-    "fuelConsumptionCalc": EntityCategory.DIAGNOSTIC,
-    "ecosrvHttps": EntityCategory.DIAGNOSTIC,
-    "ecosrvAddr": EntityCategory.DIAGNOSTIC,
-    "routerType": EntityCategory.DIAGNOSTIC,
-    "ecosrvSoftVer": EntityCategory.DIAGNOSTIC,
-    "moduleEcoSTERSoftVer": EntityCategory.DIAGNOSTIC,
-}
-
-# =============================================================================
-# ENTITY VALUE LIMITS
-# =============================================================================
-ENTITY_MIN_VALUE = {
-    "tempCOSet": 27,
-    "tempCWUSet": 20,
-}
-
-ENTITY_MAX_VALUE = {
-    "tempCOSet": 68,
-    "tempCWUSet": 55,
-}
-
-ENTITY_STEP = {
-    "tempCOSet": 1,
-    "tempCWUSet": 1,
-}
-
-# Sensor value mappings for both display and icon support
-# Note: mode and transmission use the same OPERATION_MODE_NAMES mapping
-# to ensure consistent state display across both sensors
-SENSOR_MODE_MAPPING: dict[int, str] = OPERATION_MODE_NAMES
-
-SENSOR_LAMBDA_STATUS_MAPPING: dict[int, str] = {
-    0: "stop",
-    1: "start",
-    2: "working",
-    3: "error",
-    4: "maintenance",
-}
-
-SENSOR_STATUS_CWU_MAPPING: dict[int, str] = {
-    0: "not_set",
-    1: "set",
-    128: "no_temp_set",
-}
-
-SENSOR_STATUS_CO_MAPPING: dict[int, str] = {
-    0: "off",  # modeTurnOff - TURNED OFF
-    1: "stop",  # modeStop - STOP
-    2: "fire_up",  # modeKindle - FIRE UP / KINDLING
-    3: "operation",  # modeWork - OPERATION / WORK
-    4: "supervision",  # modeSupervision - SUPERVISION
-    5: "paused",  # modeHalt - HALTED / PAUSED
-    6: "cleaning",  # modeCleaning - CLEANING
-    7: "burning_off",  # modeExtinction - BURNING OFF / EXTINCTION
-    8: "alarm",  # modeAlarm - ALARM
-    9: "manual",  # modeManual - MANUAL
-    10: "unsealing",  # modeUnsealing - UNSEALING
-    11: "other",  # modeOther - OTHER
-    12: "stabilization",  # modeStabilization - STABILIZATION
-    13: "purge",  # modePurge - PURGE
-    14: "check_flame",  # modeCheckFlame - CHECK FLAME
-    15: "flame_losing",  # modeFlameLosing - FLAME LOSING
-    16: "prevention",  # modePrevention - PREVENTION
-    17: "work_grate",  # modeWorkGrate - WORK GRATE
-    18: "supervision_grate",  # modeSupervisionGrate - SUPERVISION GRATE
-    19: "calibration",  # modeCalibration - CALIBRATION
-    20: "maintain",  # modeMaintain - MAINTAIN / MAINTENANCE
-    21: "afterburning",  # modeAfterburning - AFTERBURNING
-    22: "chimney_sweep",  # modeChimneySwep - CHIMNEY SWEEP
-    23: "heating",  # modeHeats - HEATING
-    24: "open_door",  # modeOpenDoor - OPEN DOOR
-    25: "cooling",  # modeCooling - COOLING
-    26: "safe",  # modeSafe - SAFE MODE
-}
-
-SENSOR_THERMOSTAT_MAPPING: dict[int, str] = {
-    0: "off",
-    1: "on",
-}
-
-# ecoMAX360i-specific value mappings
-SENSOR_FLAP_VALVE_STATES_MAPPING: dict[int, str] = {
-    0: "central_heating",
-    3: "domestic_hot_water",
-}
-
-SENSOR_HEAT_DEMANDED_MAPPING: dict[int, str] = {
-    0: "heat",
-    1: "off",
-}
-
-SENSOR_WATER_PUMP_RUNNING_MAPPING: dict[int, str] = {
-    0: "on",
-    1: "unknown",
-    2: "off",
-}
-
-# =============================================================================
-# ENUM SENSOR OPTIONS
-# =============================================================================
-# Options for SensorDeviceClass.ENUM sensors - displayed in HA Developer Tools
-SENSOR_ENUM_OPTIONS: dict[str, list[str]] = {
-    "mode": [*OPERATION_MODE_NAMES.values(), STATE_UNKNOWN],
-    "transmission": [*OPERATION_MODE_NAMES.values(), STATE_UNKNOWN],
-    "statusCO": [*SENSOR_STATUS_CO_MAPPING.values(), STATE_UNKNOWN],
-    # ecoMAX360i enum options
-    "flapValveStates": [*SENSOR_FLAP_VALVE_STATES_MAPPING.values(), STATE_UNKNOWN],
-    "HeatDemanded": [*SENSOR_HEAT_DEMANDED_MAPPING.values(), STATE_UNKNOWN],
-    "WaterPumpRunning": list(
-        dict.fromkeys([*SENSOR_WATER_PUMP_RUNNING_MAPPING.values(), STATE_UNKNOWN])
-    ),
-}
-
-# =============================================================================
-# SCHEDULE CONSTANTS (ecoMAX "em" protocol)
-# =============================================================================
-# User-friendly name -> API key mapping for ecomaxSchedules
-SCHEDULE_TYPE_MAP: dict[str, str] = {
-    "boiler": "boilerTZ",
-    "boiler_clean": "boilerCleanTZ",
-    "boiler_work": "boilerWorkTZ",
-    "circulation_pump": "circPumpTZ",
-    "exchanger_clean": "exchangerCleanTZ",
-    "mixer_1": "mixer1TZ",
-    "mixer_2": "mixer2TZ",
-    "mixer_3": "mixer3TZ",
-    "mixer_4": "mixer4TZ",
-    "mixer_5": "mixer5TZ",
-    "mixer_6": "mixer6TZ",
-    "mixer_7": "mixer7TZ",
-    "mixer_8": "mixer8TZ",
-    "mixer_9": "mixer9TZ",
-    "mixer_10": "mixer10TZ",
-    "thermostat_1": "thermostat1TZ",
-    "thermostat_2": "thermostat2TZ",
-    "thermostat_3": "thermostat3TZ",
-    "water_heater": "cwuTZ",
-    "water_heater_2": "cwu2TZ",
-}
-
-# Reverse map: API key -> user-friendly name
-SCHEDULE_TYPE_REVERSE_MAP: dict[str, str] = {v: k for k, v in SCHEDULE_TYPE_MAP.items()}
-
-# Day ordering matches the ecoNET web interface (index 0 = Sunday)
-SCHEDULE_WEEKDAYS: list[str] = [
-    "sunday",
-    "monday",
-    "tuesday",
-    "wednesday",
-    "thursday",
-    "friday",
-    "saturday",
-]
