+++ custom_components/econet300/const/entity_maps.py
@@ -1,0 +1,781 @@
+"""Per-entity attribute maps: units, state classes, device classes, precision, category and value limits."""
+
+from homeassistant.components.binary_sensor import (
+    BinarySensorDeviceClass as _BinarySensorDeviceClass,
+)
+from homeassistant.components.number import NumberDeviceClass as _NumberDeviceClass
+from homeassistant.components.sensor import (
+    SensorDeviceClass as _SensorDeviceClass,
+    SensorStateClass as _SensorStateClass,
+)
+from homeassistant.const import (
+    PERCENTAGE as _PERCENTAGE,
+    SIGNAL_STRENGTH_DECIBELS_MILLIWATT as _SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
+    EntityCategory as _EntityCategory,
+    UnitOfEnergy as _UnitOfEnergy,
+    UnitOfPower as _UnitOfPower,
+    UnitOfTemperature as _UnitOfTemperature,
+    UnitOfTime as _UnitOfTime,
+)
+
+from .core import SENSOR_FUEL_STREAM
+
+# =============================================================================
+# ENTITY UNIT MAPPINGS
+# =============================================================================
+# By default all sensors unit_of_measurement are None
+ENTITY_UNIT_MAP = {
+    "tempCO": _UnitOfTemperature.CELSIUS,
+    "tempCOSet": _UnitOfTemperature.CELSIUS,
+    "tempCWUSet": _UnitOfTemperature.CELSIUS,
+    "tempExternalSensor": _UnitOfTemperature.CELSIUS,
+    "tempFeeder": _UnitOfTemperature.CELSIUS,
+    "lambdaLevel": _PERCENTAGE,
+    "lambdaSet": _PERCENTAGE,
+    "workAt100": _UnitOfTime.HOURS,
+    "workAt50": _UnitOfTime.HOURS,
+    "workAt30": _UnitOfTime.HOURS,
+    "FeederWork": _UnitOfTime.HOURS,
+    "thermoTemp": _UnitOfTemperature.CELSIUS,
+    "fanPower": _PERCENTAGE,
+    "tempFlueGas": _UnitOfTemperature.CELSIUS,
+    "mixerSetTemp1": _UnitOfTemperature.CELSIUS,
+    "mixerSetTemp2": _UnitOfTemperature.CELSIUS,
+    "mixerSetTemp3": _UnitOfTemperature.CELSIUS,
+    "mixerSetTemp4": _UnitOfTemperature.CELSIUS,
+    "mixerSetTemp5": _UnitOfTemperature.CELSIUS,
+    "mixerSetTemp6": _UnitOfTemperature.CELSIUS,
+    "mixerTemp1": _UnitOfTemperature.CELSIUS,
+    "mixerTemp2": _UnitOfTemperature.CELSIUS,
+    "mixerTemp3": _UnitOfTemperature.CELSIUS,
+    "mixerTemp4": _UnitOfTemperature.CELSIUS,
+    "mixerTemp5": _UnitOfTemperature.CELSIUS,
+    "mixerTemp6": _UnitOfTemperature.CELSIUS,
+    "tempBack": _UnitOfTemperature.CELSIUS,
+    "tempCWU": _UnitOfTemperature.CELSIUS,
+    "boilerPower": _PERCENTAGE,
+    "boilerPowerKW": _UnitOfPower.KILO_WATT,
+    "fuelLevel": _PERCENTAGE,
+    "tempUpperBuffer": _UnitOfTemperature.CELSIUS,
+    "tempLowerBuffer": _UnitOfTemperature.CELSIUS,
+    "signal": _SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
+    "quality": _PERCENTAGE,
+    "valveMixer1": _PERCENTAGE,
+    "burnerOutput": _PERCENTAGE,
+    "mixerTemp": _UnitOfTemperature.CELSIUS,
+    "mixerSetTemp": _UnitOfTemperature.CELSIUS,
+    # ecoMAX850R2-X specific units
+    "fuelConsum": _PERCENTAGE,
+    SENSOR_FUEL_STREAM: "kg/h",  # Mass flow rate (confirmed from ecoNET cloud API)
+    "transmission": None,
+    # ecoSTER thermostat units
+    "ecoSterTemp1": _UnitOfTemperature.CELSIUS,
+    "ecoSterTemp2": _UnitOfTemperature.CELSIUS,
+    "ecoSterTemp3": _UnitOfTemperature.CELSIUS,
+    "ecoSterTemp4": _UnitOfTemperature.CELSIUS,
+    "ecoSterTemp5": _UnitOfTemperature.CELSIUS,
+    "ecoSterTemp6": _UnitOfTemperature.CELSIUS,
+    "ecoSterTemp7": _UnitOfTemperature.CELSIUS,
+    "ecoSterTemp8": _UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp1": _UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp2": _UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp3": _UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp4": _UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp5": _UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp6": _UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp7": _UnitOfTemperature.CELSIUS,
+    "ecoSterSetTemp8": _UnitOfTemperature.CELSIUS,
+    "ecoSterMode1": None,
+    "ecoSterMode2": None,
+    "ecoSterMode3": None,
+    "ecoSterMode4": None,
+    "ecoSterMode5": None,
+    "ecoSterMode6": None,
+    "ecoSterMode7": None,
+    "ecoSterMode8": None,
+    # ecoMAX360i
+    "Circuit2thermostatTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit3thermostatTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit4thermostatTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit5thermostatTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit6thermostatTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit7thermostatTemp": _UnitOfTemperature.CELSIUS,
+    "TempClutch": _UnitOfTemperature.CELSIUS,
+    "TempWthr": _UnitOfTemperature.CELSIUS,
+    "TempCircuit2": _UnitOfTemperature.CELSIUS,
+    "TempCircuit3": _UnitOfTemperature.CELSIUS,
+    "TempCircuit4": _UnitOfTemperature.CELSIUS,
+    "TempCircuit5": _UnitOfTemperature.CELSIUS,
+    "TempCircuit6": _UnitOfTemperature.CELSIUS,
+    "TempCircuit7": _UnitOfTemperature.CELSIUS,
+    "TempBuforUp": _UnitOfTemperature.CELSIUS,
+    "TempCWU": _UnitOfTemperature.CELSIUS,
+    "TempBuforDown": _UnitOfTemperature.CELSIUS,
+    "heatingUpperTemp": _UnitOfTemperature.CELSIUS,
+    # Circuit1thermostat is a thermostat state (no "Temp" suffix), not a temperature.
+    "Circuit1thermostat": None,
+    "flapValveStates": None,
+    "HeatDemanded": None,
+    "WaterPumpRunning": None,
+    # Heat pump (Axen) units
+    "AxenOutdoorTemp": _UnitOfTemperature.CELSIUS,
+    "AxenOutgoingTemp": _UnitOfTemperature.CELSIUS,
+    "AxenReturnTemp": _UnitOfTemperature.CELSIUS,
+    "AxenCompressorFreq": "Hz",
+    "AxenUpperPump": None,
+    "AxenWorkState": None,
+    "HeatSourceCalcPresetTemp": _UnitOfTemperature.CELSIUS,
+    # Heat pump compressor temperature units
+    "afterCompressorTemp": _UnitOfTemperature.CELSIUS,
+    "beforeCompressorTemp": _UnitOfTemperature.CELSIUS,
+    "exhaustGasTemp": _UnitOfTemperature.CELSIUS,
+    "outdoorTemp": _UnitOfTemperature.CELSIUS,
+    # Heat pump status units
+    "HPStatusWorkMode": None,
+    "HPStatusPresetTemp": _UnitOfTemperature.CELSIUS,
+    "HPStatusControl": None,
+    # SSA (weather compensation) units
+    "ssaState": None,
+    "ssaCorr": _UnitOfTemperature.CELSIUS,
+    "ssaPrevTemp": _UnitOfTemperature.CELSIUS,
+    # Circuit comfort/eco temperature setpoints
+    "Circuit1ComfortTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit1EcoTemp": _UnitOfTemperature.CELSIUS,
+    "BufferTargetTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit2ComfortTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit2EcoTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit3ComfortTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit3EcoTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit4ComfortTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit4EcoTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit5ComfortTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit5EcoTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit6ComfortTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit6EcoTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit7ComfortTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit7EcoTemp": _UnitOfTemperature.CELSIUS,
+    # ecoMAX360i informationParams / editParams sensors
+    "TargetFlowTemp": _UnitOfTemperature.CELSIUS,
+    "ActualFlowTemp": _UnitOfTemperature.CELSIUS,
+    "ActualReturnTemp": _UnitOfTemperature.CELSIUS,
+    "FanSpeed": "rpm",
+    "HeatPumpAmbient": _UnitOfTemperature.CELSIUS,
+    "ActualDHWTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit1DesiredLWT": _UnitOfTemperature.CELSIUS,
+    "ElectricalPower": _UnitOfPower.KILO_WATT,
+    "ThermalPower": _UnitOfPower.KILO_WATT,
+    "COP": None,
+    "SCOP": None,
+    "FlowRate": "L/min",
+    "AXENREGISTER64": None,
+    "AXENREGISTER65": None,
+    # Phoenix / extended heat-pump units (decomposed from EXTRA_SENSORS)
+    "BuforCalcSetTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit1CalcTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit2CalcTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit4CalcTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit5CalcTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit6CalcTemp": _UnitOfTemperature.CELSIUS,
+    "Circuit7CalcTemp": _UnitOfTemperature.CELSIUS,
+    "HDWsetpointcalculate": _UnitOfTemperature.CELSIUS,
+    "CoolingPower": _UnitOfPower.KILO_WATT,
+    "HeatingPower": _UnitOfPower.KILO_WATT,
+    "ElectricPower": _UnitOfPower.KILO_WATT,
+    "currentFlow": "m³/h",
+    "PHNXcoil_temp": _UnitOfTemperature.CELSIUS,
+    "PHNXdischarge_temp": _UnitOfTemperature.CELSIUS,
+    "PHNXinletTemp": _UnitOfTemperature.CELSIUS,
+    "PHNXinletTemp_pump": _UnitOfTemperature.CELSIUS,
+    "PHNXinletTemp_pump2": _UnitOfTemperature.CELSIUS,
+    "PHNXinletTemp_pump3": _UnitOfTemperature.CELSIUS,
+    "PHNXinletTemp_pump5": _UnitOfTemperature.CELSIUS,
+    "PHNXoutletTemp": _UnitOfTemperature.CELSIUS,
+    "PHNXoutletTemp_pump2": _UnitOfTemperature.CELSIUS,
+    "PHNXoutletTemp_pump3": _UnitOfTemperature.CELSIUS,
+    "PHNXoutletTemp_pump4": _UnitOfTemperature.CELSIUS,
+    "PHNXoutletTemp_pump5": _UnitOfTemperature.CELSIUS,
+    "PHNXreg2045": _UnitOfTemperature.CELSIUS,
+    "PHNXreg2046": _UnitOfTemperature.CELSIUS,
+    "PHNXreg2071": "Hz",
+    "PHNXreg2074": "rpm",
+    "PHNXsuctionTemp": _UnitOfTemperature.CELSIUS,
+    # ecoSOL specific units (ecoSOL [n] models)
+    "T1": _UnitOfTemperature.CELSIUS,
+    "T2": _UnitOfTemperature.CELSIUS,
+    "T3": _UnitOfTemperature.CELSIUS,
+    "T4": _UnitOfTemperature.CELSIUS,
+    "T5": _UnitOfTemperature.CELSIUS,
+    "T6": _UnitOfTemperature.CELSIUS,
+    "TzCWU": _UnitOfTemperature.CELSIUS,
+    "P1": None,
+    "P2": None,
+    "H": None,
+    "Moc_chwilowa": _UnitOfPower.KILO_WATT,
+    "Uzysk_ca_kowity": _UnitOfEnergy.KILO_WATT_HOUR,
+    # ecoSOL diagnostic units
+    "ecosrvAddr": None,
+    "softVer": None,
+    "routerType": None,
+    "protocolType": None,
+    "controllerID": None,
+    "ecosrvSoftVer": None,
+    "moduleEcoSTERSoftVer": None,
+}
+
+# =============================================================================
+# ENTITY STATE CLASS MAPPINGS
+# =============================================================================
+# By default all sensors state_class are MEASUREMENT
+STATE_CLASS_MAP: dict[str, _SensorStateClass | None] = {
+    "lambdaStatus": None,
+    "mode": None,
+    "thermostat": None,
+    "statusCO": None,
+    "statusCWU": None,
+    "softVer": None,
+    "controllerID": None,
+    "moduleASoftVer": None,
+    "moduleBSoftVer": None,
+    "moduleCSoftVer": None,
+    "moduleLambdaSoftVer": None,
+    "modulePanelSoftVer": None,
+    # Diagnostic sensors (non-numeric)
+    "routerType": None,
+    "protocolType": None,
+    "moduleEcoSTERSoftVer": None,
+    "ecosrvSoftVer": None,
+    "transmission": None,
+    "ecoSterMode1": None,
+    "ecoSterMode2": None,
+    "ecoSterMode3": None,
+    "ecoSterMode4": None,
+    "ecoSterMode5": None,
+    "ecoSterMode6": None,
+    "ecoSterMode7": None,
+    "ecoSterMode8": None,
+    "P1": None,
+    "P2": None,
+    "H": None,
+    "Moc_chwilowa": _SensorStateClass.MEASUREMENT,
+    "Uzysk_ca_kowity": _SensorStateClass.TOTAL_INCREASING,
+    # ecoMAX360i
+    "PS": None,
+    "heating_work_state_pump4": None,
+    "flapValveStates": None,
+    "HeatDemanded": None,
+    "WaterPumpRunning": None,
+    # ecoMAX360i informationParams / editParams sensors
+    "FanSpeed": None,
+    "COP": None,
+    "SCOP": None,
+    "FlowRate": None,
+    "AXENREGISTER64": None,
+    "AXENREGISTER65": None,
+    # ecoMAX360i heat pump status sensors (not measurements)
+    "HPStatusWorkMode": None,
+    "HPStatusControl": None,
+    "AxenUpperPump": None,
+    "AxenWorkState": None,
+    "ssaState": None,
+    # Phoenix / extended heat-pump status sensors (not measurements)
+    "Circuit1thermostat": None,
+    "HPStatusBuffHeatStat": None,
+    "HPStatusCircPStat0": None,
+    "HPStatusCircPStat1": None,
+    "HPStatusCircPStat2": None,
+    "HPStatusCircPStat3": None,
+    "HPStatusCircPStat4": None,
+    "HPStatusCircPStat5": None,
+    "HPStatusCircPStat6": None,
+    "HPStatusFlowHeatStat": None,
+    "HPStatusHdwHeatStat": None,
+    "HPStatusUhsStat": None,
+    "HP_work_state_set_pump2": None,
+    "HP_work_state_set_pump3": None,
+    "HP_work_state_set_pump4": None,
+    "HP_work_state_set_pump5": None,
+    # Phoenix / extended heat-pump measurements (numeric; explicit because the
+    # default state_class is None to avoid bogus statistics on all-sensors keys).
+    "BuforCalcSetTemp": _SensorStateClass.MEASUREMENT,
+    "Circuit1CalcTemp": _SensorStateClass.MEASUREMENT,
+    "Circuit2CalcTemp": _SensorStateClass.MEASUREMENT,
+    "Circuit4CalcTemp": _SensorStateClass.MEASUREMENT,
+    "Circuit5CalcTemp": _SensorStateClass.MEASUREMENT,
+    "Circuit6CalcTemp": _SensorStateClass.MEASUREMENT,
+    "Circuit7CalcTemp": _SensorStateClass.MEASUREMENT,
+    "Circuit2thermostatTemp": _SensorStateClass.MEASUREMENT,
+    "Circuit3thermostatTemp": _SensorStateClass.MEASUREMENT,
+    "HDWsetpointcalculate": _SensorStateClass.MEASUREMENT,
+    "HPStatusPresetTemp": _SensorStateClass.MEASUREMENT,
+    "CoolingPower": _SensorStateClass.MEASUREMENT,
+    "HeatingPower": _SensorStateClass.MEASUREMENT,
+    "ElectricPower": _SensorStateClass.MEASUREMENT,
+    "currentFlow": _SensorStateClass.MEASUREMENT,
+    "TempBuforDown": _SensorStateClass.MEASUREMENT,
+    "TempBuforUp": _SensorStateClass.MEASUREMENT,
+    "TempCWU": _SensorStateClass.MEASUREMENT,
+    "TempCircuit2": _SensorStateClass.MEASUREMENT,
+    "TempCircuit3": _SensorStateClass.MEASUREMENT,
+    "TempCircuit4": _SensorStateClass.MEASUREMENT,
+    "TempCircuit5": _SensorStateClass.MEASUREMENT,
+    "TempCircuit6": _SensorStateClass.MEASUREMENT,
+    "TempCircuit7": _SensorStateClass.MEASUREMENT,
+    "TempClutch": _SensorStateClass.MEASUREMENT,
+    "TempWthr": _SensorStateClass.MEASUREMENT,
+    "PHNXcoil_temp": _SensorStateClass.MEASUREMENT,
+    "PHNXdischarge_temp": _SensorStateClass.MEASUREMENT,
+    "PHNXinletTemp": _SensorStateClass.MEASUREMENT,
+    "PHNXinletTemp_pump": _SensorStateClass.MEASUREMENT,
+    "PHNXinletTemp_pump2": _SensorStateClass.MEASUREMENT,
+    "PHNXinletTemp_pump3": _SensorStateClass.MEASUREMENT,
+    "PHNXinletTemp_pump5": _SensorStateClass.MEASUREMENT,
+    "PHNXoutletTemp": _SensorStateClass.MEASUREMENT,
+    "PHNXoutletTemp_pump2": _SensorStateClass.MEASUREMENT,
+    "PHNXoutletTemp_pump3": _SensorStateClass.MEASUREMENT,
+    "PHNXoutletTemp_pump4": _SensorStateClass.MEASUREMENT,
+    "PHNXoutletTemp_pump5": _SensorStateClass.MEASUREMENT,
+    "PHNXreg2045": _SensorStateClass.MEASUREMENT,
+    "PHNXreg2046": _SensorStateClass.MEASUREMENT,
+    "PHNXreg2071": _SensorStateClass.MEASUREMENT,
+    "PHNXreg2074": _SensorStateClass.MEASUREMENT,
+    "PHNXsuctionTemp": _SensorStateClass.MEASUREMENT,
+}
+
+# =============================================================================
+# ENTITY DEVICE CLASS MAPPINGS
+# =============================================================================
+
+ENTITY_SENSOR_DEVICE_CLASS_MAP: dict[str, _SensorDeviceClass | None] = {
+    "tempFeeder": _SensorDeviceClass.TEMPERATURE,
+    "tempExternalSensor": _SensorDeviceClass.TEMPERATURE,
+    "tempCO": _SensorDeviceClass.TEMPERATURE,
+    "tempCOSet": _SensorDeviceClass.TEMPERATURE,
+    "boilerPower": _SensorDeviceClass.POWER_FACTOR,
+    "boilerPowerKW": _SensorDeviceClass.POWER,
+    "fanPower": _SensorDeviceClass.POWER_FACTOR,
+    "tempFlueGas": _SensorDeviceClass.TEMPERATURE,
+    "mixerSetTemp1": _SensorDeviceClass.TEMPERATURE,
+    "mixerSetTemp2": _SensorDeviceClass.TEMPERATURE,
+    "mixerSetTemp3": _SensorDeviceClass.TEMPERATURE,
+    "mixerSetTemp4": _SensorDeviceClass.TEMPERATURE,
+    "mixerSetTemp5": _SensorDeviceClass.TEMPERATURE,
+    "mixerSetTemp6": _SensorDeviceClass.TEMPERATURE,
+    "mixerTemp1": _SensorDeviceClass.TEMPERATURE,
+    "mixerTemp2": _SensorDeviceClass.TEMPERATURE,
+    "mixerTemp3": _SensorDeviceClass.TEMPERATURE,
+    "mixerTemp4": _SensorDeviceClass.TEMPERATURE,
+    "mixerTemp5": _SensorDeviceClass.TEMPERATURE,
+    "mixerTemp6": _SensorDeviceClass.TEMPERATURE,
+    "mixerTemp": _SensorDeviceClass.TEMPERATURE,
+    "mixerSetTemp": _SensorDeviceClass.TEMPERATURE,
+    "tempBack": _SensorDeviceClass.TEMPERATURE,
+    "tempCWU": _SensorDeviceClass.TEMPERATURE,
+    "statusCO": _SensorDeviceClass.ENUM,
+    "statusCWU": None,
+    "tempUpperBuffer": _SensorDeviceClass.TEMPERATURE,
+    "tempLowerBuffer": _SensorDeviceClass.TEMPERATURE,
+    "signal": _SensorDeviceClass.SIGNAL_STRENGTH,
+    "servoMixer1": _SensorDeviceClass.ENUM,
+    "mode": _SensorDeviceClass.ENUM,
+    "transmission": _SensorDeviceClass.ENUM,
+    # ecoMAX850R2-X specific device classes
+    "fuelConsum": _SensorDeviceClass.POWER_FACTOR,
+    # fuelStream: No device class - HA has no SensorDeviceClass for mass flow rate (kg/h)
+    # ecoSTER thermostat device classes
+    "ecoSterTemp1": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp2": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp3": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp4": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp5": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp6": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp7": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterTemp8": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp1": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp2": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp3": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp4": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp5": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp6": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp7": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterSetTemp8": _SensorDeviceClass.TEMPERATURE,
+    "ecoSterMode1": None,
+    "ecoSterMode2": None,
+    "ecoSterMode3": None,
+    "ecoSterMode4": None,
+    "ecoSterMode5": None,
+    "ecoSterMode6": None,
+    "ecoSterMode7": None,
+    "ecoSterMode8": None,
+    # ecoMAX360i
+    "Circuit2thermostatTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit3thermostatTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit4thermostatTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit5thermostatTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit6thermostatTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit7thermostatTemp": _SensorDeviceClass.TEMPERATURE,
+    "TempClutch": _SensorDeviceClass.TEMPERATURE,
+    "TempWthr": _SensorDeviceClass.TEMPERATURE,
+    "TempCircuit2": _SensorDeviceClass.TEMPERATURE,
+    "TempCircuit3": _SensorDeviceClass.TEMPERATURE,
+    "TempCircuit4": _SensorDeviceClass.TEMPERATURE,
+    "TempCircuit5": _SensorDeviceClass.TEMPERATURE,
+    "TempCircuit6": _SensorDeviceClass.TEMPERATURE,
+    "TempCircuit7": _SensorDeviceClass.TEMPERATURE,
+    "TempBuforUp": _SensorDeviceClass.TEMPERATURE,
+    "TempCWU": _SensorDeviceClass.TEMPERATURE,
+    "TempBuforDown": _SensorDeviceClass.TEMPERATURE,
+    "heatingUpperTemp": _SensorDeviceClass.TEMPERATURE,
+    # Circuit1thermostat is a thermostat state (no "Temp" suffix), not a temperature.
+    "Circuit1thermostat": None,
+    "flapValveStates": _SensorDeviceClass.ENUM,
+    "HeatDemanded": _SensorDeviceClass.ENUM,
+    "WaterPumpRunning": _SensorDeviceClass.ENUM,
+    # Heat pump (Axen) device classes
+    "AxenOutdoorTemp": _SensorDeviceClass.TEMPERATURE,
+    "AxenOutgoingTemp": _SensorDeviceClass.TEMPERATURE,
+    "AxenReturnTemp": _SensorDeviceClass.TEMPERATURE,
+    "AxenCompressorFreq": _SensorDeviceClass.FREQUENCY,
+    "AxenUpperPump": None,
+    "AxenWorkState": None,
+    "HeatSourceCalcPresetTemp": _SensorDeviceClass.TEMPERATURE,
+    # Heat pump compressor temperature device classes
+    "afterCompressorTemp": _SensorDeviceClass.TEMPERATURE,
+    "beforeCompressorTemp": _SensorDeviceClass.TEMPERATURE,
+    "exhaustGasTemp": _SensorDeviceClass.TEMPERATURE,
+    "outdoorTemp": _SensorDeviceClass.TEMPERATURE,
+    # Heat pump status device classes
+    "HPStatusWorkMode": None,
+    "HPStatusPresetTemp": _SensorDeviceClass.TEMPERATURE,
+    "HPStatusControl": None,
+    # SSA (weather compensation) device classes
+    "ssaState": None,
+    "ssaCorr": _SensorDeviceClass.TEMPERATURE,
+    "ssaPrevTemp": _SensorDeviceClass.TEMPERATURE,
+    # Circuit comfort/eco temperature device classes
+    "Circuit1ComfortTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit1EcoTemp": _SensorDeviceClass.TEMPERATURE,
+    "BufferTargetTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit2ComfortTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit2EcoTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit3ComfortTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit3EcoTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit4ComfortTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit4EcoTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit5ComfortTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit5EcoTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit6ComfortTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit6EcoTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit7ComfortTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit7EcoTemp": _SensorDeviceClass.TEMPERATURE,
+    # ecoMAX360i informationParams / editParams device classes
+    "TargetFlowTemp": _SensorDeviceClass.TEMPERATURE,
+    "ActualFlowTemp": _SensorDeviceClass.TEMPERATURE,
+    "ActualReturnTemp": _SensorDeviceClass.TEMPERATURE,
+    "FanSpeed": None,
+    "HeatPumpAmbient": _SensorDeviceClass.TEMPERATURE,
+    "ActualDHWTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit1DesiredLWT": _SensorDeviceClass.TEMPERATURE,
+    "ElectricalPower": _SensorDeviceClass.POWER,
+    "ThermalPower": _SensorDeviceClass.POWER,
+    "COP": _SensorDeviceClass.POWER_FACTOR,
+    "SCOP": _SensorDeviceClass.POWER_FACTOR,
+    "FlowRate": None,
+    "AXENREGISTER64": None,
+    "AXENREGISTER65": None,
+    # Phoenix / extended heat-pump device classes (decomposed from EXTRA_SENSORS)
+    "BuforCalcSetTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit1CalcTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit2CalcTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit4CalcTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit5CalcTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit6CalcTemp": _SensorDeviceClass.TEMPERATURE,
+    "Circuit7CalcTemp": _SensorDeviceClass.TEMPERATURE,
+    "HDWsetpointcalculate": _SensorDeviceClass.TEMPERATURE,
+    "CoolingPower": _SensorDeviceClass.POWER,
+    "HeatingPower": _SensorDeviceClass.POWER,
+    "ElectricPower": _SensorDeviceClass.POWER,
+    "PHNXcoil_temp": _SensorDeviceClass.TEMPERATURE,
+    "PHNXdischarge_temp": _SensorDeviceClass.TEMPERATURE,
+    "PHNXinletTemp": _SensorDeviceClass.TEMPERATURE,
+    "PHNXinletTemp_pump": _SensorDeviceClass.TEMPERATURE,
+    "PHNXinletTemp_pump2": _SensorDeviceClass.TEMPERATURE,
+    "PHNXinletTemp_pump3": _SensorDeviceClass.TEMPERATURE,
+    "PHNXinletTemp_pump5": _SensorDeviceClass.TEMPERATURE,
+    "PHNXoutletTemp": _SensorDeviceClass.TEMPERATURE,
+    "PHNXoutletTemp_pump2": _SensorDeviceClass.TEMPERATURE,
+    "PHNXoutletTemp_pump3": _SensorDeviceClass.TEMPERATURE,
+    "PHNXoutletTemp_pump4": _SensorDeviceClass.TEMPERATURE,
+    "PHNXoutletTemp_pump5": _SensorDeviceClass.TEMPERATURE,
+    "PHNXreg2045": _SensorDeviceClass.TEMPERATURE,
+    "PHNXreg2046": _SensorDeviceClass.TEMPERATURE,
+    "PHNXreg2071": _SensorDeviceClass.FREQUENCY,
+    "PHNXsuctionTemp": _SensorDeviceClass.TEMPERATURE,
+    # ecoSOL specific device classes (ecoSOL [n] models)
+    "T1": _SensorDeviceClass.TEMPERATURE,
+    "T2": _SensorDeviceClass.TEMPERATURE,
+    "T3": _SensorDeviceClass.TEMPERATURE,
+    "T4": _SensorDeviceClass.TEMPERATURE,
+    "T5": _SensorDeviceClass.TEMPERATURE,
+    "T6": _SensorDeviceClass.TEMPERATURE,
+    "TzCWU": _SensorDeviceClass.TEMPERATURE,
+    "P1": None,
+    "P2": None,
+    "H": None,
+    "Moc_chwilowa": _SensorDeviceClass.POWER,
+    "Uzysk_ca_kowity": _SensorDeviceClass.ENERGY,
+    # ecoSOL diagnostic device classes
+    "ecosrvAddr": None,
+    "softVer": None,
+    "routerType": None,
+    "protocolType": None,
+    "controllerID": None,
+    "ecosrvSoftVer": None,
+}
+
+# Number entity device classes
+ENTITY_NUMBER_SENSOR_DEVICE_CLASS_MAP = {
+    "tempCOSet": _NumberDeviceClass.TEMPERATURE,
+    "tempCWUSet": _NumberDeviceClass.TEMPERATURE,
+    # ecoMAX850R2-X specific number entities
+    "mixerSetTemp5": _NumberDeviceClass.TEMPERATURE,
+}
+
+# Binary sensor device classes
+ENTITY_BINARY_DEVICE_CLASS_MAP = {
+    # By default all binary sensors device_class are RUNNING
+    "mainSrv": _BinarySensorDeviceClass.CONNECTIVITY,
+    "wifi": _BinarySensorDeviceClass.CONNECTIVITY,
+    "lan": _BinarySensorDeviceClass.CONNECTIVITY,
+    "fuelConsumptionCalc": _BinarySensorDeviceClass.RUNNING,
+    "ecosrvHttps": _BinarySensorDeviceClass.CONNECTIVITY,
+    # ecoMAX850R2-X specific binary sensors
+    "contactGZC": _BinarySensorDeviceClass.CONNECTIVITY,
+    "contactGZCActive": _BinarySensorDeviceClass.CONNECTIVITY,
+    # Alarm sensors
+    "alarmActive": _BinarySensorDeviceClass.PROBLEM,
+}
+
+# =============================================================================
+# ENTITY PRECISION MAPPINGS
+# =============================================================================
+# Add only keys where precision more than 0 needed
+ENTITY_PRECISION = {
+    "tempFeeder": 1,
+    "tempExternalSensor": 1,
+    "lambdaLevel": 1,
+    "lambdaSet": 1,
+    "tempCO": 1,
+    "mixerTemp1": 1,
+    "mixerTemp2": 1,
+    "mixerTemp3": 1,
+    "mixerTemp4": 1,
+    "mixerTemp5": 1,
+    "mixerTemp6": 1,
+    "tempBack": 2,
+    "tempUpperBuffer": 1,
+    "tempLowerBuffer": 1,
+    "tempCWU": 1,
+    "tempFlueGas": 1,
+    "fanPower": 0,
+    "statusCO": None,
+    "statusCWU": None,
+    "thermostat": None,
+    "lambdaStatus": None,
+    "mode": None,
+    "transmission": None,
+    "softVer": None,
+    "controllerID": None,
+    "moduleASoftVer": None,
+    "moduleBSoftVer": None,
+    "moduleCSoftVer": None,
+    "moduleLambdaSoftVer": None,
+    "modulePanelSoftVer": None,
+    "moduleEcoSTERSoftVer": None,
+    "ecosrvSoftVer": None,
+    # ecoMAX360i
+    "PS": None,
+    "Circuit2thermostatTemp": 1,
+    "Circuit3thermostatTemp": 1,
+    "Circuit4thermostatTemp": 1,
+    "Circuit5thermostatTemp": 1,
+    "Circuit6thermostatTemp": 1,
+    "Circuit7thermostatTemp": 1,
+    "TempClutch": 1,
+    "TempWthr": 1,
+    "TempCircuit2": 1,
+    "TempCircuit3": 1,
+    "TempCircuit4": 1,
+    "TempCircuit5": 1,
+    "TempCircuit6": 1,
+    "TempCircuit7": 1,
+    "TempBuforUp": 1,
+    "TempCWU": 1,
+    "TempBuforDown": 1,
+    "flapValveStates": None,
+    "HeatDemanded": None,
+    "WaterPumpRunning": None,
+    # Heat pump (Axen) precision
+    "AxenOutdoorTemp": 1,
+    "AxenOutgoingTemp": 1,
+    "AxenReturnTemp": 1,
+    "AxenCompressorFreq": 0,
+    # Circuit comfort/eco temperature precision
+    "Circuit1ComfortTemp": 1,
+    "Circuit1EcoTemp": 1,
+    "BufferTargetTemp": 1,
+    "Circuit2ComfortTemp": 1,
+    "Circuit2EcoTemp": 1,
+    "Circuit3ComfortTemp": 1,
+    "Circuit3EcoTemp": 1,
+    "Circuit4ComfortTemp": 1,
+    "Circuit4EcoTemp": 1,
+    "Circuit5ComfortTemp": 1,
+    "Circuit5EcoTemp": 1,
+    "Circuit6ComfortTemp": 1,
+    "Circuit6EcoTemp": 1,
+    "Circuit7ComfortTemp": 1,
+    "Circuit7EcoTemp": 1,
+    # ecoMAX360i informationParams / editParams precision
+    "TargetFlowTemp": 1,
+    "ActualFlowTemp": 1,
+    "ActualReturnTemp": 1,
+    "FanSpeed": 0,
+    "HeatPumpAmbient": 1,
+    "ActualDHWTemp": 1,
+    "Circuit1DesiredLWT": 1,
+    "ElectricalPower": 2,
+    "ThermalPower": 2,
+    "COP": 2,
+    "SCOP": 2,
+    "FlowRate": 1,
+    "AXENREGISTER64": 1,
+    "AXENREGISTER65": 0,
+    # Heat pump compressor temperature precision
+    "afterCompressorTemp": 1,
+    "beforeCompressorTemp": 1,
+    "exhaustGasTemp": 1,
+    "outdoorTemp": 1,
+    # Heat pump status precision
+    "HPStatusWorkMode": None,
+    "HPStatusPresetTemp": 1,
+    "HPStatusControl": None,
+    "AxenUpperPump": None,
+    "AxenWorkState": None,
+    "HeatSourceCalcPresetTemp": 0,
+    # SSA (weather compensation) precision
+    "ssaState": None,
+    "ssaCorr": 0,
+    "ssaPrevTemp": 1,
+    # ecoSTER thermostat precision
+    "ecoSterTemp1": 1,
+    "ecoSterTemp2": 1,
+    "ecoSterTemp3": 1,
+    "ecoSterTemp4": 1,
+    "ecoSterTemp5": 1,
+    "ecoSterTemp6": 1,
+    "ecoSterTemp7": 1,
+    "ecoSterTemp8": 1,
+    "ecoSterSetTemp1": 1,
+    "ecoSterSetTemp2": 1,
+    "ecoSterSetTemp3": 1,
+    "ecoSterSetTemp4": 1,
+    "ecoSterSetTemp5": 1,
+    "ecoSterSetTemp6": 1,
+    "ecoSterSetTemp7": 1,
+    "ecoSterSetTemp8": 1,
+    "heatingUpperTemp": 1,
+    # Circuit1thermostat is a thermostat state (no "Temp" suffix), not a temperature.
+    "Circuit1thermostat": None,
+    "heating_work_state_pump4": None,
+    # ecoSOL specific precision (ecoSOL [n] models)
+    "T1": 1,
+    "T2": 1,
+    "T3": 1,
+    "T4": 1,
+    "T5": 1,
+    "T6": 1,
+    "TzCWU": 1,
+    "P1": None,
+    "P2": None,
+    "H": None,
+    "Moc_chwilowa": 2,
+    "Uzysk_ca_kowity": 2,
+    # ecoSOL diagnostic precision
+    "ecosrvAddr": None,
+    "routerType": None,
+    "protocolType": None,
+    # Phoenix / extended heat-pump precision (decomposed from EXTRA_SENSORS)
+    "BuforCalcSetTemp": 1,
+    "Circuit1CalcTemp": 1,
+    "Circuit2CalcTemp": 1,
+    "Circuit4CalcTemp": 1,
+    "Circuit5CalcTemp": 1,
+    "Circuit6CalcTemp": 1,
+    "Circuit7CalcTemp": 1,
+    "HDWsetpointcalculate": 0,
+    "CoolingPower": 2,
+    "HeatingPower": 2,
+    "ElectricPower": 2,
+    "currentFlow": 2,
+    "PHNXcoil_temp": 1,
+    "PHNXdischarge_temp": 1,
+    "PHNXinletTemp": 1,
+    "PHNXinletTemp_pump": 1,
+    "PHNXinletTemp_pump2": 1,
+    "PHNXinletTemp_pump3": 1,
+    "PHNXinletTemp_pump5": 1,
+    "PHNXoutletTemp": 1,
+    "PHNXoutletTemp_pump2": 1,
+    "PHNXoutletTemp_pump3": 1,
+    "PHNXoutletTemp_pump4": 1,
+    "PHNXoutletTemp_pump5": 1,
+    "PHNXreg2045": 1,
+    "PHNXreg2046": 1,
+    "PHNXreg2071": 1,
+    "PHNXreg2074": 0,
+    "PHNXsuctionTemp": 1,
+}
+
+# =============================================================================
+# ENTITY CATEGORY MAPPINGS
+# =============================================================================
+ENTITY_CATEGORY = {
+    "signal": _EntityCategory.DIAGNOSTIC,
+    "quality": _EntityCategory.DIAGNOSTIC,
+    "softVer": _EntityCategory.DIAGNOSTIC,
+    "moduleASoftVer": _EntityCategory.DIAGNOSTIC,
+    "moduleBSoftVer": _EntityCategory.DIAGNOSTIC,
+    "modulePanelSoftVer": _EntityCategory.DIAGNOSTIC,
+    "moduleLambdaSoftVer": _EntityCategory.DIAGNOSTIC,
+    "protocolType": _EntityCategory.DIAGNOSTIC,
+    "controllerID": _EntityCategory.DIAGNOSTIC,
+    "moduleCSoftVer": _EntityCategory.DIAGNOSTIC,
+    "mainSrv": _EntityCategory.DIAGNOSTIC,
+    "wifi": _EntityCategory.DIAGNOSTIC,
+    "lan": _EntityCategory.DIAGNOSTIC,
+    "fuelConsumptionCalc": _EntityCategory.DIAGNOSTIC,
+    "ecosrvHttps": _EntityCategory.DIAGNOSTIC,
+    "ecosrvAddr": _EntityCategory.DIAGNOSTIC,
+    "routerType": _EntityCategory.DIAGNOSTIC,
+    "ecosrvSoftVer": _EntityCategory.DIAGNOSTIC,
+    "moduleEcoSTERSoftVer": _EntityCategory.DIAGNOSTIC,
+}
+
+# =============================================================================
+# ENTITY VALUE LIMITS
+# =============================================================================
+ENTITY_MIN_VALUE = {
+    "tempCOSet": 27,
+    "tempCWUSet": 20,
+}
+
+ENTITY_MAX_VALUE = {
+    "tempCOSet": 68,
+    "tempCWUSet": 55,
+}
+
+ENTITY_STEP = {
+    "tempCOSet": 1,
+    "tempCWUSet": 1,
+}
+
